% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.r
\name{addIntrDB}
\alias{addIntrDB}
\title{Add interaction database to CytoSignal object}
\usage{
addIntrDB(
  object,
  gene_to_uniprot,
  intr.db.diff_dep,
  intr.db.cont_dep,
  inter.index
)
}
\arguments{
\item{object}{CytoSignal object}

\item{gene_to_uniprot}{df, gene to uniprot symbols mapping}

\item{intr.db.diff_dep}{intr.db for diffusable ligands + non-diffusable receptors}

\item{intr.db.cont_dep}{intr.db for contact dependent ligands + receptors}

\item{inter.index}{df, collection of intraction information}
}
\value{
a CytoSignal object
}
\description{
Add interaction database to CytoSignal object
}
