% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.r
\name{addVelo}
\alias{addVelo}
\title{Add velocity data to CytoSignal object}
\usage{
addVelo(object, velo.s, velo.u)
}
\arguments{
\item{object}{CytoSignal object}

\item{velo.s}{matrix of spliced velo}

\item{velo.u}{matrix of unspliced velo}
}
\description{
Add velocity data to CytoSignal object
}
