% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{changeUniprot.dgCMatrix}
\alias{changeUniprot.dgCMatrix}
\title{Sub function for changeUniprot, input a sparse matrix}
\usage{
\method{changeUniprot}{dgCMatrix}(dge.raw, gene_to_uniprot, verbose = T)
}
\arguments{
\item{dge.raw}{A sparse matrix}

\item{gene_to_uniprot}{A data frame with gene name and uniprot id}

\item{verbose}{Whether to print out the progress}
}
\value{
A sparse matrix
}
\description{
Sub function for changeUniprot, input a sparse matrix
}
