% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.r
\name{changeUniprot}
\alias{changeUniprot}
\alias{changeUniprot.matrix_like}
\alias{changeUniprot.CytoSignal}
\title{Subset gene expression matrix according to availability in the UNIPROT
database.}
\usage{
changeUniprot(object, ...)

\method{changeUniprot}{matrix_like}(object, gene_to_uniprot, verbose = TRUE, ...)

\method{changeUniprot}{CytoSignal}(object, verbose = TRUE, ...)
}
\arguments{
\item{object}{A CytoSignal object or a matrix.}

\item{...}{Parameters passed to S3 methods.}

\item{gene_to_uniprot}{Please use \code{\link{g_to_u}}.}

\item{verbose}{Whether to show information of the progress.}
}
\value{
CytoSignal object with subset matrix updated in \code{counts} slot,
or the subset matrix.
}
\description{
Subset genes to those available in \code{\link{g_to_u}}, the gene name to
UNIPROT ID databse provided with \code{\link{addIntrDB}}
}
