% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.r
\name{createCytoSignal}
\alias{createCytoSignal}
\title{Create a CytoSignal object}
\usage{
createCytoSignal(
  raw.data,
  cells.loc,
  clusters = NULL,
  name = NULL,
  parameters = NULL,
  log = NULL,
  version = NULL
)
}
\arguments{
\item{raw.data}{raw data matrix}

\item{cells.loc}{cells location matrix}

\item{clusters}{cluster information}

\item{name}{name of the dataset}

\item{parameters}{parameters used}

\item{log}{log of the processing}

\item{version}{version of the package}
}
\value{
a CytoSignal object
}
\description{
Create a CytoSignal object
}
