% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{csColors}
\alias{csColors}
\title{CytoSignal Color Palette}
\format{
An object of class \code{character} of length 51.
}
\usage{
csColors
}
\description{
A vector of colors used in visualizing discrete values. Originally taken from
\code{ggsci::pal_igv(palette = 'default')(51)}
}
\keyword{datasets}
