% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{csdb}
\alias{csdb}
\alias{print.csdb}
\title{CytoSignal Database class}
\usage{
\method{print}{csdb}(x, ...)
}
\arguments{
\item{x}{A \code{csdb} object.}

\item{...}{See \code{tibble::\link[tibble]{formatting}}.}
}
\value{
None.
}
\description{
A tibble data frame with the following columns:
\itemize{
\item{\code{interactors}: Interaction ID in the format "partner_a-partner_b",
where \code{partner_a} and \code{partner_b} are gene names of ligand and receptor,
respectively. For complexes, gene names are joined by "+".}
\item{\code{ligands}: Gene names of ligand(s), separated by semicolon if multiple.}
\item{\code{receptors}: Gene names of receptor(s), separated by semicolon if multiple.}
\item{\code{type}: Type of interaction, either \code{'diffusion'} or \code{'contact'}.}
\item{\code{alt_name}: Alternative name for the interaction, using complex names
if applicable.}
}
Additional columns might be seen as this class is also used to deliver top
interaction ranking result from downstream analyses.
}
\seealso{
\code{\link[=subsetCSDB]{subsetCSDB()}}
}
