% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intrDEG.R
\name{csIntrDEG}
\alias{csIntrDEG}
\title{csIntrDEG - Result holder for interaction-associated DEG analysis}
\description{
An S3 class to store the results of interaction-associated DEG detection.
It is basically a list object, with class attribute \code{csIntrDEG}.

It contains the following fields:
\describe{
\item{geneSignif}{A list of data frames, each containing the DEGs
associated with an interaction. Each data frame has the following columns:
\itemize{
\item \code{gene}: Gene name.
\item \code{nCluster}: Number of clusters where the gene is significantly
up-regulated when the interaction is significant.
\item \code{meanLFC}: Mean log fold-change of the gene across the
significant clusters.
\item \code{effectSize}: Effect size metric combining mean log
fold-change and number of clusters.
}}
\item{fdrThresh}{FDR threshold used to define if an interaction is
significantly observed at each spot.}
\item{clusterVar}{A factor vector of cluster labels for each spot.}
\item{models}{Only available after \code{\link{refineIntrDEG}}.

A list of fitted elastic net regression models per interaction.}
\item{clusterSignif}{Only available after \code{\link{refineIntrDEG}}.

A list of character vectors, each containing the clusters that are selected
by the elastic net regression model per interaction.}
}
}
\seealso{
\code{\link{findIntrDEG}}, \code{\link{refineIntrDEG}}, \code{\link{plotIntrDEGHeatmap}}
}
