% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{csToSCE}
\alias{csToSCE}
\alias{SCEToCS}
\title{Conversion between CytoSignal object and SingleCellExperiment object}
\usage{
csToSCE(object)

SCEToCS(object, coords, assay = "counts", clusters = NULL)
}
\arguments{
\item{object}{A CytoSignal object or a Seurat object.}

\item{coords}{Name of the \code{reducedDim} in the SCE object to get spatial
coordinates. Required.}

\item{assay}{Name of a SCE assay to extract raw count matrix from. Default
\code{"counts"}}

\item{clusters}{Name of the metadata variable in the SCE object
\code{colData} that contains the cluster identities of each spot. Default
\code{NULL} does not pull anything out.}
}
\value{
\itemize{
\item {
\code{SCEToCS} returns a CytoSignal object that includes only the
raw counts, coordinates, and cluster identities of the spatial assay.
}

\item {
\code{csToSCE} returns a SingleCellExperiment object. The main experiment
stores the transcriptomics data with the following information:
\itemize{
\item{\code{assay(sce, 'counts')} - The raw counts.}
\item{\code{colData(sce)} - The metadata table for spots, only containing
column \code{'clusters'}.}
\item{\code{reducedDim(sce, 'spatial')} - The spatial coordinates.}
}
The CytoSignal analysis results are mainly stored in alternative experiments.
For example, for the inference made with GauEps-Raw model will be stored in
\code{altExp(sce, 'GauEps-Raw')}, with the following information:
\itemize{
\item{\code{assay(altExp(sce, 'GauEps-Raw'), 'lrscore')} - The LR-score
matrix per diffusion-dependent interaction by spots.}
\item{\code{assay(altExp(sce, 'GauEps-Raw'), 'isSignif.spx')} - A binary
sparse matrix of the same dimensionality, where a 1-valued entry indicates
that the interaction is significant at the spot.}
\item{\code{rowData(altExp(sce, 'GauEps-Raw'))} - The interaction metadata
table. Column \code{'intr.name'} stores the interpretable interaction
names. The other columns stores the ranking of the interactions at
different significance inferrence levels.}
\item{\code{reducedDim(altExp(sce, 'GauEps-Raw'), 'spatial')} - A redundant
copy of the spatial coordinates, for convenience.}
}
}
}
}
\description{
Converts a CytoSignal object to a SingleCellExperiment (SCE) object or from a
SCE object to a CytoSignal object.
}
