% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{csToSeurat}
\alias{csToSeurat}
\alias{SeuratToCS}
\title{Conversion between CytoSignal object and Seurat object}
\usage{
csToSeurat(object)

SeuratToCS(
  object,
  layer = "counts",
  assay = "Spatial",
  image = NULL,
  cluster = NULL
)
}
\arguments{
\item{object}{A CytoSignal object or a Seurat object.}

\item{layer}{Name of Seurat V5 layer name to extract the raw count matrix
from the assay. Default \code{"counts"}.}

\item{assay}{Name of a Seurat assay to extract count matrix from or put into.
Default \code{"Spatial"}}

\item{image}{Name of the spatial image object in the Seurat object where the
spatial coordinates are stored. Default \code{NULL} uses the first image
available to the specified assay.}

\item{cluster}{Name of the metadata variable in the Seurat object that
contains the cluster identities of each spot. Default \code{NULL} uses the
activate identity in the Seurat object.}
}
\value{
\itemize{
\item {
\code{SeuratToCS} returns a CytoSignal object that includes only the
raw counts, coordinates, and cluster identities of the spatial assay.
}
\item {
\code{csToSeurat} returns a Seurat object with "Spatial" assay that contains
the raw counts. When LR-score and signifcance inferrence results are present,
another assay called "CytoSignal" is created to store the LR-score matrices
and binary matrices indicating whether an interaction is significant at a
spot. The spatial coordinates are coerced to a "SlideSeq" object regardless
of the exact technology being imported from. Interpretable interaction names,
rather than the CPI IDs of each interaction, are available as feature
metadata of the CytoSignal assay.
}
}
}
\description{
Converts a CytoSignal object to a Seurat object or from a Seurat object to a
CytoSignal object.
}
