% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.r
\docType{class}
\name{CytoSignal-class}
\alias{CytoSignal-class}
\alias{CytoSignal}
\title{The CytoSignal Class}
\description{
The CytoSignal object is created from one ST dataset. To construct a
CytoSignal object, the user needs to provide both DGE and cell locations. The class provides functions for data
preprocessing, integrative analysis, and visualization.
}
\details{
The key slots used in the CytoSignal object are described below.
}
\section{Slots}{

\describe{
\item{\code{counts}}{raw dge, gene x cell}

\item{\code{cell.loc}}{matrix of cell locations}

\item{\code{clusters}}{cluster assignments}

\item{\code{imputation}}{list of imputation objs}

\item{\code{lrscore}}{list of lrScore objs}

\item{\code{velo}}{list of velo objs}

\item{\code{intr.valid}}{list of intr database}

\item{\code{cell.data}}{data frame of cell data}

\item{\code{parameters}}{List of parameters for gaussian imputation}

\item{\code{log}}{Log of each main steps}

\item{\code{version}}{Version of package used to create object}
}}

