% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object2.R
\docType{class}
\name{CytoSignal2-class}
\alias{CytoSignal2-class}
\alias{createCytoSignal2}
\alias{show,cytosignal2-method}
\alias{.DollarNames.cytosignal2}
\alias{$.cytosignal2}
\alias{$<-.cytosignal2}
\title{CytoSignal2-class}
\usage{
createCytoSignal2(rawData, spatial, cluster = NULL, ...)

\S4method{show}{cytosignal2}(object)

\method{.DollarNames}{cytosignal2}(x, pattern = "")

\method{$}{cytosignal2}(x, name)

\method{$}{cytosignal2}(x, name) <- value
}
\arguments{
\item{rawData}{A sparse matrix of gene expression data with genes as rows and
spots as columns.}

\item{spatial}{The spatial coordinates coercible to a matrix, with rows as
spots and two columns for x and y.}

\item{cluster}{A vector/factor of cluster assignments for each spot. The
length must equal to the number of spots (columns) in \code{rawData}.}

\item{...}{Additional arguments directly passed to object slots.}

\item{object, x}{The \code{cytosignal2} object.}

\item{pattern}{A character string containing a regular expression to match column names of metadata.}

\item{name}{The name of the metadata column to access.}

\item{value}{The value to assign.}
}
\description{
Container object for CytoSignal analysis that holds input data, parameters,
and results.
}
\section{Slots}{

\describe{
\item{\code{rawData}}{A sparse matrix of gene expression data with genes as rows and spots as columns.}

\item{\code{spatial}}{A matrix of spatial coordinates with rows as spots and two columns (x and y).}

\item{\code{metadata}}{A tibble (data.frame) containing metadata for each spot.}

\item{\code{intrDB}}{An interaction database object of class \code{csdb}. See \link{CytoSignal-database} for how to construct.}

\item{\code{parameters}}{A list of parameters used in the analysis.}

\item{\code{neighborDiff}}{A sparse matrix representing the diffusion-dependent neighborhood model.}

\item{\code{neighborCont}}{A sparse matrix representing the contact-dependent neighborhood model.}

\item{\code{LRScore}}{A sparse matrix of ligand-receptor interaction scores.}

\item{\code{nullECDF}}{A sparse matrix of empirical cumulative distribution functions
for interaction scores. A permutation p-value can be derived by subtracting
these values from 1.}

\item{\code{significance}}{A list containing significance results for the interaction scores.}

\item{\code{version}}{A package_version object indicating the version of CytoSignal
used to create the object.}
}}

\examples{
spatial <- matrix(runif(100), ncol = 2)
rownames(spatial) <- paste0('Spot', 1:50)
cs <- createCytoSignal2(
   rawData = Matrix::rsparsematrix(
       nrow = 100, ncol = 50, density = 0.1,
       dimnames = list(
           paste0('Gene', 1:100),
           paste0('Spot', 1:50)
       )
   ),
   spatial = spatial
)
cs$total_counts
cs$cluster <- sample(letters[1:3], 50, replace = TRUE)
}
