% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{db.diff}
\alias{db.diff}
\alias{db.cont}
\alias{inter.index}
\alias{g_to_u}
\title{Interaction database derived from CellphoneDB V2}
\format{
\code{db.diff} and \code{db.cont} are lists of factor objects.
\code{db.diff} is for diffusion dependent interactions and \code{db.cont} is
for contact dependent interactions. Both lists have the following three
entries of the same format:
\itemize{
\item{\code{$ligands} - A factor object where the values are IDs of
interactions and names are UNIPROT IDs of the ligand protein components that
match to the interaction.}
\item{\code{$receptors} - A factor object of the same format as
\code{$ligands} but for the receptor part of the interactions.}
\item{\code{$combined} - The concatenation of the two factors above.}
}

\code{inter.index} is a data.frame object with the meta-information of each
interaction.

\code{g_to_u} is a data.frame object that map gene names to UNIPROT ID of
their protein product.

An object of class \code{list} of length 3.

An object of class \code{data.frame} with 1396 rows and 11 columns.

An object of class \code{data.frame} with 977 rows and 3 columns.
}
\source{
CellphoneDB V2
}
\usage{
db.diff

db.cont

inter.index

g_to_u
}
\description{
The referencing database that contains the list of curated interactions, the
meta-information of each interaction, the ligand and receptor of each
interaction, the protein components of both part of each interaction. The
protein components are presented with UNIPROT ID, so we also provide
\code{\link{g_to_u}} that maps UNIPROT ID to gene names.
}
\keyword{datasets}
