% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nebula.R
\name{dim.mergedCytoSignal}
\alias{dim.mergedCytoSignal}
\title{dim Method for mergedCytoSignal object}
\usage{
\method{dim}{mergedCytoSignal}(x)
}
\arguments{
\item{x}{A mergedCytoSignal object}
}
\value{
A numeric vector with the number of spots/cells and NA
}
\description{
Only returns nrow(x) which indicates the number of spots/cells
}
