% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRvelo.r
\name{.inferVeloLR.matrix_like}
\alias{.inferVeloLR.matrix_like}
\title{#' Compute the LR velo for specific ligand-receptor imputation obj pairs
#'
#' @param object A Cytosignal object
#' @param lig.slot The ligand slot to use
#' @param recep.slot The receptor slot to use
#' @param intr.db.name The intr database name to use
#' @param nn.use The neighbor index as niche
#'
#' @return A Cytosignal object
#' @export
#'
inferVeloLR <- function(
object,
...
) {
UseMethod(generic = 'inferVeloLR', object = object)
}}
\usage{
.inferVeloLR.matrix_like(
  dge.lig,
  dge.recep,
  dge.lig.velo,
  dge.recep.velo,
  lig.fac,
  recep.fac
)
}
\description{
#' Compute the LR velo for specific ligand-receptor imputation obj pairs
#'
#' @param object A Cytosignal object
#' @param lig.slot The ligand slot to use
#' @param recep.slot The receptor slot to use
#' @param intr.db.name The intr database name to use
#' @param nn.use The neighbor index as niche
#'
#' @return A Cytosignal object
#' @export
#'
inferVeloLR <- function(
object,
...
) {
UseMethod(generic = 'inferVeloLR', object = object)
}
}
