% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nebula.R
\name{$.mergedCytoSignal}
\alias{$.mergedCytoSignal}
\alias{$<-.mergedCytoSignal}
\title{Extract metadata variable from mergedCytoSignal object}
\usage{
\method{$}{mergedCytoSignal}(x, name)

\method{$}{mergedCytoSignal}(x, name) <- value
}
\description{
Fast expression for getting and setting metadata variables in a
\linkS4class{mergedCytoSignal} object.
}
