% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intrDEG.R
\name{findIntrDEG}
\alias{findIntrDEG}
\title{Test gene expression association with interaction significance}
\usage{
findIntrDEG(
  object,
  interaction = NULL,
  intrFdrThresh = 0.05,
  minExp = 100,
  minSpot = 100,
  degFdrThresh = 0.05
)
}
\arguments{
\item{object}{A \code{\linkS4class{cytosignal2}} object.}

\item{interaction}{Which interactions to use. Users can provide a vector of
interaction selection (name or indices applicable to rows of \code{intrDB(object)}).
Or a subset of the database itself (e.g. \code{getTopIntr(object, topN = 2, type = 'cont')})}

\item{intrFdrThresh}{FDR threshold to claim that an interaction is
significantly observed in a spot. Default \code{0.05}.}

\item{minExp, minSpot}{Quality control parameters for top interaction
selection if \code{interaction} is not provided. See \code{\link{getTopIntr}}.}

\item{degFdrThresh}{FDR threshold to claim that a gene is differentially
expressed in the Wilcoxon's rank-sum test. Default \code{0.05}.}

\item{topN}{Maximum number of top DEGs to append per interaction per cluster.
Default \code{30}.}
}
\value{
A list of character vectors, each containing the DEGs associated
with the significance of an interaction. If only one interaction is tested,
a single character vector is returned.
}
\description{
This function performs a one-sided Wilcoxon's rank-sum test on each gene,
between the cells that receive significant interaction signal and the cells
that do not. The test is done per interaction and per cluster (if available).
For each interaction, the function collects the genes that show significant
up-regulation when spots receive significant interaction signal. Genes are
eventually ranked by an effect size metric that combines the average log
fold-change and the number of clusters where the gene is significantly
up-regulated.
}
\seealso{
\code{\link{refineIntrDEG}}, \code{\link{plotIntrDEGHeatmap}}
}
