% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor.R
\name{findNeighbor}
\alias{findNeighbor}
\alias{findNeighborGauEB2}
\alias{findNeighborDT2}
\title{Build weighted neighborhood graph that models ligand received by spots}
\usage{
findNeighbor(object, ncores = 1L)

findNeighborGauEB2(object, ncores = 1L)

findNeighborDT2(object)
}
\arguments{
\item{object}{A \code{cytosignal2} object. With \code{\link{setParams}} called.}

\item{ncores}{A positive Integer specifying the number of threads to use.
This is only effective for finding neighbors with the Gaussian Epsilon ball
method, and when OpenMP is supported in the system when compiling the
package. Default \code{1L} does not use parallel computing.}
}
\value{
A \code{cytosignal2} object with \code{neighborDiff} and/or \code{neighborCont}
slots filled with a sparse square matrix representing the neighborhood graph.
}
\description{
\code{findNeighbor()} is a wrapper function that calls
\code{findNeighborGauEB2()} and \code{findNeighborDT2()}.

Parameters \bold{must} be set with \code{\link{setParams}} \bold{before}
running these functions, in order to appropriately scale the spatial
coordinates and assign the distance limits.

\code{findNeighborGauEB2()} uses Gaussian Epsilon ball method to detect
neighbors within the limited distance and assign weights based on a Gaussian
kernel. This models the diffusion-dependent ligand-receptor interactions.

\code{findNeighborDT2()} uses Delaunay triangulation to detect neighbors that
are directly contacting each other within the limited distance and assigns
equal weights. This models the contact-dependent ligand-receptor
interactions.
}
\seealso{
\code{\link[=setParams]{setParams()}}

\code{\link[=plotNeighbor]{plotNeighbor()}}
}
