% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{findNN}
\alias{findNN}
\title{Identify nearest neighbors for each location using different strategies for
different types of interactions}
\usage{
findNN(
  object,
  eps = NULL,
  sigma = NULL,
  self.weight = "auto",
  dt.mode = "weight_sum_2"
)
}
\arguments{
\item{object}{A \linkS4class{CytoSignal} object.}

\item{eps}{Gaussian Epsilon Ball method parameter. A numeric scalar. The
radius of the Gaussian Epsilon ball. Default \code{NULL} uses parameter
inferred with \code{\link{inferEpsParams}}.}

\item{sigma}{Gaussian Epsilon Ball method parameter. The \eqn{\sigma} of the
Gaussian kernel. Default \code{NULL} uses parameter inferred with
\code{\link{inferEpsParams}}.}

\item{self.weight}{Gaussian Epsilon Ball method parameter. Weight of the
index cell. Use a number between 0-1 or a string \code{"auto"} or
\code{"sum_1"}. Default \code{"auto"}.}

\item{dt.mode}{Delaunay Triangulation method parameter. A string to select
weight normalization strategy. Default \code{"weight_sum_2"} normalizes
neighbor weights and keep self-weights as 1, alternatively
\code{"weight_sum_1"} normalizes all weights.}

\item{max.r}{Delaunay Triangulation method parameter. A numeric scalar for
the maximum radius of the edges. Default \code{NULL} uses parameter inferred
with \code{\link{inferEpsParams}}.}
}
\value{
A \linkS4class{CytoSignal} object updated. \code{object@imputation}
slot will be updated with three new entries: \code{object@imputation$GauEps},
\code{object@imputation$DT} and \code{object@imputation$Raw}.
}
\description{
This is a wrapper function of \code{\link{findNNGauEB}},
\code{\link{findNNDT}} and \code{\link{findNNRaw}}. We use a Gaussian Epsilon
ball to identify the nearest neighbors that can contribute to the diffusible
ligands. And then uses a Delaunay triangulation to identify the nearest
neighbors that can contribute to the contact-dependent ligands. The
identified nearest neighbors will then be used for imputing the \eqn{L} or
\eqn{R} value for each location.
}
\examples{
\dontrun{
object <- findNN(object)
}
}
