% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{findNNDT.CytoSignal}
\alias{findNNDT.CytoSignal}
\title{Sub function for findNNDT, input a Cytosignal object}
\usage{
\method{findNNDT}{CytoSignal}(object, weight = 2, max.r = NULL)
}
\arguments{
\item{object}{A Cytosignal object}

\item{weight}{The weight of the Delaunay triangulation}

\item{max.r}{The maximum radius of the edges, by default is the r.diffuse.scale}
}
\value{
A Cytosignal object
}
\description{
Sub function for findNNDT, input a Cytosignal object
}
