% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{findNNDT.matrix}
\alias{findNNDT.matrix}
\title{Sub function for findNNDT, input a matrix}
\usage{
\method{findNNDT}{matrix}(cells.loc, weight.sum = 2)
}
\arguments{
\item{cells.loc}{A matrix of cells location}

\item{weight.sum}{The sum of the weights}
}
\value{
A list of neighbors
}
\description{
Sub function for findNNDT, input a matrix
}
