% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{findNNDT}
\alias{findNNDT}
\title{Find the direct connected neighbor of each cell, using Delaunay triangulation}
\usage{
findNNDT(object, ...)
}
\arguments{
\item{object}{A Cytosignal object}

\item{weight.sum}{The sum of the weights}
}
\value{
A Cytosignal object
}
\description{
Find the direct connected neighbor of each cell, using Delaunay triangulation
}
