% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{findNNGauEB.matrix}
\alias{findNNGauEB.matrix}
\title{Sub function for findNNGauEB, input a sparse matrix}
\usage{
\method{findNNGauEB}{matrix}(cells.loc, eps, sigma = 0.15, weight.sum = 1)
}
\arguments{
\item{eps}{The radius of the epsilon ball}

\item{sigma}{The sigma of the Gaussian kernel}

\item{weight.sum}{The sum of the weights}

\item{object}{A Cytosignal object}
}
\value{
A Cytosignal object
}
\description{
Sub function for findNNGauEB, input a sparse matrix
}
