% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{findNNRaw}
\alias{findNNRaw}
\title{Create a ImpData object using raw data without imputation}
\usage{
findNNRaw(object)
}
\arguments{
\item{object}{A CytoSignal object}
}
\value{
Returns the original object with result updated in "imputation" slot.
An "ImpData" object is created with its "nn.graph" taken from the result of
\code{\link{findNNDT}}.
}
\description{
This function do not attempt to find neighbors but simply creates internal
data structure which is helpful for downstream analysis.
}
