% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_gene.R
\name{format.CytosignalIntrDEG}
\alias{format.CytosignalIntrDEG}
\title{Format a CytosignalIntrDEG object to string}
\usage{
\method{format}{CytosignalIntrDEG}(x, ...)
}
\arguments{
\item{x}{A \code{CytosignalIntrDEG} object.}

\item{...}{Passed to other methods}
}
\value{
A string representation of the object
}
\description{
Format a CytosignalIntrDEG object to string
}
