% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{generalParam}
\alias{generalParam}
\title{Basic arguments for convenience documentation}
\usage{
generalParam(object, interaction, fdrThresh = 0.05)
}
\arguments{
\item{object}{A \code{\linkS4class{cytosignal2}} object.}

\item{interaction}{Which interactions to use. Users can provide a vector of
interaction selection (name or indices applicable to rows of \code{intrDB(object)}).
Or a subset of the database itself (e.g. \code{getTopIntr(object, topN = 2, type = 'cont')})}

\item{fdrThresh}{FDR threshold to claim that an interaction is significantly
observed in a spot. Default \code{0.05}.}
}
\value{
Invisible NULL value.
}
\description{
This documentation is not for users.
}
