% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{getCellPhoneDB}
\alias{getCellPhoneDB}
\title{Download and make DB from public CellPhoneDB data}
\usage{
getCellPhoneDB(
  path = "https://github.com/ventolab/cellphonedb-data/archive/refs/tags/v5.0.0.zip",
  saveLoc = paste0(tempdir(), "_cellphonedb"),
  interactionInput = "data/interaction_input.csv",
  proteinInput = "data/protein_input.csv",
  complexInput = "data/complex_input.csv",
  geneInput = "data/gene_input.csv"
)
}
\arguments{
\item{path}{URL or local path to CellPhoneDB-data folder.}

\item{saveLoc}{Local path to save downloaded CellPhoneDB data if downloading
from URL. Defaults to a temporary directory.}

\item{interactionInput}{Relative path to interaction input CSV file within
the CellPhoneDB data folder.}

\item{proteinInput}{Relative path to protein input CSV file within
the CellPhoneDB data folder.}

\item{complexInput}{Relative path to complex input CSV file within
the CellPhoneDB data folder.}

\item{geneInput}{Relative path to gene input CSV file within
the CellPhoneDB data folder.}
}
\value{
A \code{\link{csdb-class}} object.
}
\description{
Download and make DB from public CellPhoneDB data
}
