% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{graphNicheLR.dgCMatrix}
\alias{graphNicheLR.dgCMatrix}
\title{Sub function for graphNicheLR, input a sparse matrix}
\usage{
\method{graphNicheLR}{dgCMatrix}(dge.lig, dge.recep, nb.id.fac, lig.fac, recep.fac)
}
\arguments{
\item{dge.lig}{A sparse matrix for ligand}

\item{dge.recep}{A sparse matrix for receptor}

\item{nb.id.fac}{A factor of neighbor indices}

\item{lig.fac}{A factor of ligand indices}

\item{recep.fac}{A factor of receptor indices}
}
\value{
A sparse matrix
}
\description{
Sub function for graphNicheLR, input a sparse matrix
}
