% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utility.r
\name{hex_coord}
\alias{hex_coord}
\title{Generate hexagon coordinates.}
\usage{
hex_coord(x, y, width, height, size = 1)
}
\arguments{
\item{x}{horizontal position of center}

\item{y}{vertical position of center}

\item{width}{hex width}

\item{height}{hex height}
}
\value{
A two column matrix with 7 times as many rows as input.
}
\description{
Long axis is horizontal. Edges clock-wise from far-left, separated by
row of missing values.
}
\examples{
x <- runif(1000)
y <- runif(1000)
res <- unique(hex_pos(x, y, 0.5, 0.5))
hexes <- hex_coord(res[, 1], res[, 2], 0.6, 0.5)

hexes <- hex_coord(res[, 1], res[, 2], rnorm(1000,.5,.3), rnorm(1000,.5,.3))

plot(hexes, type = "n")
polygon(hexes)
points(res)
}
\keyword{internal}
