% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.r
\docType{class}
\name{ImpData-class}
\alias{ImpData-class}
\alias{ImpData}
\title{The ImpData Class}
\description{
The ImpData object is created from one ST dataset. User could choose a preferred imputation method
and the class stores the imputed data, the imputed normalized data, the intr database, the intr database.
}
\details{
The key slots used in the ImpData object are described below.
}
\section{Slots}{

\describe{
\item{\code{method}}{imputation method}

\item{\code{imp.data}}{imputed data}

\item{\code{imp.norm.data}}{imputed normalized data}

\item{\code{intr.data}}{imputed normalized data subsetted by intr database}

\item{\code{intr.data.null}}{permuted imputed normalized data subsetted by intr database}

\item{\code{nn.id}}{nearest neighbor id}

\item{\code{nn.dist}}{nearest neighbor distance}

\item{\code{log}}{Log of each main steps}
}}

