% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{imputeNiche.dgCMatrix}
\alias{imputeNiche.dgCMatrix}
\title{Sub function for imputeNiche, input a sparse matrix}
\usage{
\method{imputeNiche}{dgCMatrix}(
  dge.raw,
  nb.id.fac,
  nb.dist.fac,
  weights = c("mean", "counts", "dist", "none")
)
}
\arguments{
\item{dge.raw}{A sparse matrix}

\item{nb.id.fac}{A factor of neighbors}

\item{nb.dist.fac}{A factor of weights}

\item{weights}{The weight of the Delaunay triangulation}
}
\value{
A sparse matrix, N x N graph
}
\description{
Sub function for imputeNiche, input a sparse matrix
}
