% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{imputeNiche}
\alias{imputeNiche}
\alias{imputeNiche.dgCMatrix}
\alias{imputeNiche.CytoSignal}
\title{Impute the data}
\usage{
imputeNiche(object, weights = c("mean", "counts", "dist", "none"), ...)

\method{imputeNiche}{dgCMatrix}(
  object,
  nb.id.fac,
  nb.dist.fac,
  weights = c("mean", "counts", "dist", "none"),
  ...
)

\method{imputeNiche}{CytoSignal}(
  object,
  nn.type = NULL,
  weights = c("mean", "counts", "dist", "none"),
  ...
)
}
\arguments{
\item{object}{A Cytosignal object or a dgCMatrix object of raw gene
expression matrix}

\item{weights}{The weight of the Delaunay triangulation. Choose from
\code{"mean"}, \code{"counts"}, \code{"dist"} or \code{"none"}.}

\item{...}{Arguments passed to other S3 methods}

\item{nb.id.fac}{A factor of neighbors}

\item{nb.dist.fac}{A factor of weights}

\item{nn.type}{The type of neighbors. The \code{tag} used when finding
nearest neighbors in the upstream step. If not customized, use
\code{"GauEps"} for NN found with \code{\link{findNNGauEB}}, or use
\code{"DT"} for NN found with \code{\link{findNNDT}}. Default use the most
lastly computed NN.}
}
\value{
For CytoSignal object, the original object with imputation result
updated. For dgCMatrix object, a sparse N x N graph presented as another
dgCMatrix object.
}
\description{
Impute the data
}
