% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRvelo.r
\name{imputeNicheVelo.CytoSignal}
\alias{imputeNicheVelo.CytoSignal}
\title{Sub function for imputeNicheVelo, input a Cytosignal object}
\usage{
\method{imputeNicheVelo}{CytoSignal}(object, nn.type = NULL)
}
\arguments{
\item{object}{A Cytosignal object}

\item{nn.type}{The type of neighbors}
}
\value{
A Cytosignal object
}
\description{
Sub function for imputeNicheVelo, input a Cytosignal object
}
