% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{imputeNicheVelo}
\alias{imputeNicheVelo}
\title{Impute the velocity mtx using the specified method}
\usage{
imputeNicheVelo(object, ...)
}
\arguments{
\item{object}{A Cytosignal object}

\item{...}{Other parameters}

\item{method}{The method to use for imputation}
}
\value{
A Cytosignal object
}
\description{
Impute the velocity mtx using the specified method
}
