% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{imputeVeloLR}
\alias{imputeVeloLR}
\title{Impute time derivative of \eqn{L} or \eqn{R} from the nearest neighbors of
each location}
\usage{
imputeVeloLR(object)
}
\arguments{
\item{object}{A \linkS4class{CytoSignal} object, with \code{\link{findNN}}
already run and velocity information added with \code{\link{addVelo}}.}
}
\value{
A \linkS4class{CytoSignal} object updated. Entries in
\code{object@imputation} slot will be updated with the imputation values.
}
\description{
After running \code{\link{findNN}}, we can impute the temporal \eqn{L} or
\eqn{R} change from the nearest neighbors of each location basing on the
types of nearest neighbors.
}
\examples{
\dontrun{
object <- findNN(object)
object <- imputeVeloLR(object)
}
}
