% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrscore2.R
\name{inferClusterLRScore}
\alias{inferClusterLRScore}
\title{Computes LR score with sender-receiver directionality at cluster level}
\usage{
inferClusterLRScore(object, smoothR = FALSE)
}
\arguments{
\item{object}{A \code{\linkS4class{cytosignal2}} object with neighbor graphs
constructed.}

\item{smoothR}{Logical, whether to smooth the receptor expression using the
contact-dependent neighbor graph. Default is \code{FALSE}.}
}
\value{
A 3D array of dimension (nCluster, nCluster, nInteractors), \eqn{CLR}
for cluster-wise LRscore. A value \code{CLR[s, r, i]} in the array means the
LRscore for interaction \code{i} when the sender cluster is \code{s} and the
receiver cluster is \code{r}.
dimension is the sender cluster, the second dimension is the
}
\description{
While \code{\link{inferLRScore}} computes label-free LRscores at single-spot
level, this function calculates LRscores for each pair of sender and receiver
clusters. This is done by imputing ligand amount (\eqn{L}) only from the
neighbor spots in a specific sender cluster. A cluster assignment variable
must be preset with \code{\link{setCluster}}.
}
