% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{inferCorrScore}
\alias{inferCorrScore}
\title{Infer the correspondence between LR-scores and Significance}
\usage{
inferCorrScore(object, correctBy = c("cell", "intr"), slot.use = NULL, ...)
}
\arguments{
\item{object}{A Cytosignal object}
}
\value{
A Cytosignal object
}
\description{
Min-max the LRscores, substract the significant ones, sum and average.
}
