% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrscore2.R
\name{inferECDF}
\alias{inferECDF}
\title{Calculates ECDF of null LR scores via permutation test}
\usage{
inferECDF(object, permSize = 1e+05)
}
\arguments{
\item{object}{A \code{\linkS4class{cytosignal2}} object with LR scores
inferred with \code{\link{inferLRScore}}.}

\item{permSize}{Integer, sample size to reach in the permutation test.
Default \code{1e5}.}
}
\value{
The input object with slot \code{nullECDF}
}
\description{
This function performs a permutation test to derive the empirical proportion
of null scores not exceeding the observed score, also called an empirical
cumulative distribution function (ECDF). This allows the inference of
one-sided permutation p-values for each LR interaction in each spot by
subtracting the ECDF value from 1.
}
\details{
The permutation test is performed by shuffling the spatial neighbor
relationships to impute null \eqn{L} and \eqn{R} values. The same procedure
as in \code{\link{inferLRScore}} is applied to compute null LR scores.
}
\seealso{
\code{\link[=inferLRScore]{inferLRScore()}}
}
