% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{inferEpsParams}
\alias{inferEpsParams}
\title{Infer the parameters of the Gaussian kernel}
\usage{
inferEpsParams(object, scale.factor = NULL, r.eps.real = 200, thresh = 0.001)
}
\arguments{
\item{object}{A Cytosignal object}

\item{r.eps.real}{The radius of the epsilon ball in tech resolution in um, default 200 um}

\item{thresh}{The total signal out of the epsilon ball}

\item{scale.facor}{1 spatial coord unit equals to how many µm}
}
\value{
A Cytosignal object
}
\description{
Infer the parameters of the Gaussian kernel
}
