% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_gene.R
\name{inferIntrDEG}
\alias{inferIntrDEG}
\title{Infer significant genes for each interaction}
\usage{
inferIntrDEG(
  object,
  intr = NULL,
  slot.use = NULL,
  signif.use = NULL,
  num.per.choose = 50,
  alpha.test = seq(0.5, 1, 0.1),
  seed = 1,
  minCell = 50,
  verbose = TRUE
)
}
\arguments{
\item{object}{A \code{Cytosignal} object with \code{\link{inferIntrScore}}
already run.}

\item{intr}{Specify interactions to be used. A vector of either the unique
ID of interactions or the numeric rank indices. Available IDs can be shown
with \code{showIntr(object)}. Availability of an interaction depends on the
LRscore slot to be used as well as the significance metric to be used.}

\item{slot.use}{The LRscore type to be used. See vignette for explanation.}

\item{signif.use}{The significance metric to be used. See vignette for
explanation.}

\item{num.per.choose}{The maximum number of genes to be chosen for each
wilcoxon test. A test happens for each interaction and each cell type.
Default \code{50}.}

\item{alpha.test}{A sequence of alpha values to be tested for the elastic net
regression model. The best alpha value will be chosen based on the smallest
loss. Larger alpha values result in less final selection. Default
\code{seq(0.5, 1, 0.1)}.}

\item{seed}{Random seed for controlling the cross validation. Default
\code{1}.}

\item{minCell}{Minimum number of cells where a gene is expressed to be
considered. Default \code{50}.}

\item{verbose}{Whether to print progress messages. Default \code{TRUE}.}
}
\value{
list object, each element is the result for each interaction. Each
interaction result is a list object containing the following entries:
\itemize{
\item{\code{glmnet_model} - A "cv.glmnet" object, the regression model.}
\item{\code{score_refine} - A 1-column matrix of refined LR scores of this
interaction in each cell.}
\item{\code{coef} - A matrix of coefficients of the regression model.}
\item{\code{sign_clusters} - A character vector of significant clusters in
selected by the model.}
\item{\code{sign_genes} - A character vector of significant genes selected by
the model.}
\item{\code{alpha} - The alpha value used for the final chosen model.}
\item{\code{slot.use} - The LRScore type used for this analysis.}
\item{\code{signif.use} - The significance metric used for this analysis.}
}
}
\description{
This function performs wilcoxon one-sided test for each cell type, between
cells enriched with an interaction against other cells. The top significant
genes are selected for each interaction. The expression profile of the
selected genes, together with the cell type variable, are then feeded into a
regression model for further refinement. This analysis infers the significant
genes for each interaction, and refines the LRScore using the selected genes.
}
