% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{inferScoreLR.dgCMatrix}
\alias{inferScoreLR.dgCMatrix}
\title{Sub function for inferScoreLR, input a sparse matrix}
\usage{
\method{inferScoreLR}{dgCMatrix}(dge.lig, dge.recep, lig.fac, recep.fac)
}
\arguments{
\item{dge.lig}{A sparse matrix for ligand}

\item{dge.recep}{A sparse matrix for receptor}

\item{lig.fac}{A factor of ligand indices}

\item{recep.fac}{A factor of receptor indices}
}
\value{
A sparse matrix
}
\description{
Sub function for inferScoreLR, input a sparse matrix
}
