% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{inferSignif.CytoSignal}
\alias{inferSignif.CytoSignal}
\title{Sub function for inferSignif, input is a Cytosignal object}
\usage{
\method{inferSignif}{CytoSignal}(
  object,
  p.value = 0.05,
  reads.thresh = 100,
  sig.thresh = 100,
  slot.use = NULL,
  nn.use = NULL
)
}
\arguments{
\item{object}{A Cytosignal object}

\item{p.value}{A numeric value of p-value threshold}

\item{reads.thresh}{A numeric value of reads threshold}

\item{sig.thresh}{A numeric value of significance threshold}

\item{slot.use}{A string of slot name to use}

\item{nn.use}{The lrscore obj to infer from}
}
\value{
A Cytosignal object
}
\description{
Sub function for inferSignif, input is a Cytosignal object
}
