% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{inferSignif.matrix_like}
\alias{inferSignif.matrix_like}
\title{Sub function for inferSignif, input is a sparse matrix}
\usage{
\method{inferSignif}{matrix_like}(
  dge.raw,
  lrscore.mtx,
  null.lrscore.mtx,
  correctBy = c("cell", "intr"),
  nb.fac,
  intr.db,
  gene_to_uniprot,
  p.thresh = 0.05,
  reads.thresh = 100,
  sig.thresh = 100
)
}
\arguments{
\item{dge.raw}{A sparse matrix of raw counts}

\item{lrscore.mtx}{A matrix of LR scores}

\item{null.lrscore.mtx}{A matrix of NULL LR scores}

\item{nb.fac}{A factor of nearest neighbors index}

\item{intr.db}{A interaction database}

\item{gene_to_uniprot}{A dataframe of gene to uniprot mapping}

\item{p.thresh}{A numeric value of p-value threshold}

\item{reads.thresh}{A numeric value of reads threshold}

\item{sig.thresh}{A numeric value of significance threshold}
}
\value{
A list of indexes of significant cells
}
\description{
Sub function for inferSignif, input is a sparse matrix
}
