% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{inferSpatialCorr}
\alias{inferSpatialCorr}
\title{Infer the correspondence between LR-scores and Significance}
\usage{
inferSpatialCorr(object, correctBy = c("cell", "intr"), slot.use = NULL, ...)
}
\arguments{
\item{object}{A Cytosignal object}
}
\value{
A Cytosignal object
}
\description{
Use DT neighbors to impute the p-value and LR-score for each cell, then compute the pearson
correlation between the imputed LR-score and p-value.
}
