% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrscore2.R
\name{inferSpatialVarIntr}
\alias{inferSpatialVarIntr}
\title{Infer spatial variability of LRscore with SPARK-X}
\usage{
inferSpatialVarIntr(object, nCores = 1, verbose = FALSE)
}
\arguments{
\item{object}{A \code{\linkS4class{cytosignal2}} object with LRscore
inferred with \code{\link{inferLRScore}}.}

\item{nCores}{Integer, number of threads to use. Default is \code{1}.}

\item{verbose}{Logical, whether to print detailed debugging messages from
SPARK-X.}
}
\value{
The input \code{\linkS4class{cytosignal2}} object with the following
update:
\itemize{
\item{\code{object@significance$sparkx} - A data frame containing SPARK-X
test combined p-value and BY-adjusted p-values for each interaction.}
}
}
\description{
SPARK-X builds on the covariance kernel test framework, identifying feature
quantity with spatial pattern in large scale spatial transcriptomic studies.
Here we apply SPARK-X to identify if the LRscore of each interaction shows
significant spatial variability.

Package 'SPARK' must be pre-installed with
\code{devtools::install_github(\href{https://github.com/xzhoulab/SPARK}{'xzhoulab/SPARK'})}.
}
\references{
Zhu, J., Sun, S. & Zhou, X. SPARK-X: non-parametric modeling enables scalable
and robust detection of spatial expression patterns for large spatial
transcriptomic studies. Genome Biol 22, 184 (2021).
\url{https://doi.org/10.1186/s13059-021-02404-0}

Sun, S., Zhu, J. & Zhou, X. Statistical analysis of spatial expression
patterns for spatially resolved transcriptomic studies. Nat Methods 17,
193–200 (2020). \url{https://doi.org/10.1038/s41592-019-0701-7}
}
