% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRvelo.r
\name{inferVeloLR.matrix_like}
\alias{inferVeloLR.matrix_like}
\title{Sub function for inferVeloLR, input a sparse matrix}
\usage{
\method{inferVeloLR}{matrix_like}(
  dge.lig,
  dge.recep,
  dge.lig.velo,
  dge.recep.velo,
  lig.fac,
  recep.fac
)
}
\arguments{
\item{dge.lig}{A sparse matrix for ligand}

\item{dge.recep}{A sparse matrix for receptor}

\item{lig.fac}{A factor of ligand indices}

\item{recep.fac}{A factor of receptor indices}

\item{nb.id.fac}{A factor of neighbor indices}
}
\value{
A sparse matrix
}
\description{
Sub function for inferVeloLR, input a sparse matrix
}
