% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{intrDB<-}
\alias{intrDB<-}
\title{Set interaction database for CytoSignal2 object}
\usage{
intrDB(object, toLowwords = FALSE) <- value
}
\arguments{
\item{object}{A \code{\linkS4class{cytosignal2}} object.}

\item{toLowwords}{Logical, whether to convert uppercase gene names from
database to "Lowword" format (e.g., "TP53" to "Tp53"). Default \code{FALSE}.}

\item{value}{A \code{csdb} object.}
}
\description{
Set interaction database for CytoSignal2 object
}
