% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{intrDB<-}
\alias{intrDB<-}
\alias{intrDB}
\title{Acess interaction database of a CytoSignal2 object}
\usage{
intrDB(object, toLowwords = FALSE) <- value

intrDB(
  object,
  ligandGenes = NULL,
  receptorGenes = NULL,
  type = NULL,
  LorR = FALSE
)
}
\arguments{
\item{object}{A \code{\linkS4class{cytosignal2}} object.}

\item{toLowwords}{Logical, whether to convert uppercase gene names from
database to "Lowword" format (e.g., "TP53" to "Tp53"). Default \code{FALSE}.}

\item{value}{A \code{csdb} object.}

\item{ligandGenes}{A character vector of regular expression to match ligand
genes to keep. Complexes are kept if any of their member genes match. Default
\code{NULL} skip the filtering by ligand genes.}

\item{receptorGenes}{A character vector of regular expression to match
receptor genes to keep. Complexes are kept if any of their member genes
match. Default \code{NULL} skip the filtering by receptor genes.}

\item{type}{Type of interaction to keep. Either \code{'diffusion'} or
\code{'contact'}. Default \code{NULL} skip the filtering by type.}

\item{LorR}{Logical, whether to keep interactions that match the ligand gene
pattern or the receptor gene pattern. Default \code{FALSE} keeps interactions
that match both patterns at the same time.}
}
\value{
For \code{intrDB<-}, the updated \code{\linkS4class{cytosignal2}} object with
the interaction database set.

For \code{intrDB}, a \code{csdb} object, possibly subsetted by specified
ligand/receptor genes or interaction type.
}
\description{
The setter method \code{intrDB<-} allows user to set the interaction
database for a \code{\linkS4class{cytosignal2}} object. "Lowword" case
conversion is allowed when using different species. The getter method
\code{intrDB} retrieves the interaction database in a
\code{\linkS4class{cytosignal2}} object, with optional subsetting by
ligand/receptor genes or interaction type.
}
