% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.r
\docType{class}
\name{lrScores-class}
\alias{lrScores-class}
\alias{lrScores}
\title{The lrScores Class}
\description{
The lrScores object is created from one ST dataset. User could choose two imputation methods to calculate
the ligand-receptor scores. The class stores the ligand, receptor, and interaction database, the ligand-receptor
scores, the ligand-receptor scores for permuted data, the ligand-receptor scores for permuted data, and the
log of each main steps.
}
\details{
The key slots used in the lrScores object are described below.
}
\section{Slots}{

\describe{
\item{\code{lig.slot}}{ligand database}

\item{\code{recep.slot}}{receptor database}

\item{\code{intr.slot}}{interaction database}

\item{\code{intr.list}}{list of interaction database}

\item{\code{score}}{ligand-receptor scores}

\item{\code{score.null}}{permuted ligand-receptor scores}

\item{\code{res.list}}{list of results}

\item{\code{log}}{Log of each main steps}
}}

