% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.r
\docType{class}
\name{lrVelo-class}
\alias{lrVelo-class}
\alias{lrVelo}
\title{The lrvelo Class}
\description{
The lrvelo object is created from one ST dataset. User could choose two imputation methods to calculate
the ligand-receptor scores. The class stores the index of ligand, receptor, and interaction database, inferred
lrvelo for each interaction, and the log of each main steps.
}
\details{
The key slots used in the lrVelo object are described below.
}
\section{Slots}{

\describe{
\item{\code{lig.slot}}{ligand database}

\item{\code{recep.slot}}{receptor database}

\item{\code{intr.slot}}{interaction database}

\item{\code{intr.list}}{list of interaction database}

\item{\code{velo.s}}{A matrix of spliced velo for each gene}

\item{\code{velo.u}}{A matrix of unspliced velo for each gene}

\item{\code{velo.intr}}{A sparse matrix of velo for each intr}

\item{\code{nn.id}}{A factor of nearest neighbor id. Re-do findNN since the order of cells may change!}

\item{\code{nn.dist}}{A factor of nearest neighbor distance. Re-do findNN since the order of cells may change!}

\item{\code{log}}{Log of each main steps}
}}

