% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nebula.R
\name{names.mergedCytoSignal}
\alias{names.mergedCytoSignal}
\alias{names<-.mergedCytoSignal}
\title{Dataset names of a mergedCytoSignal object}
\usage{
\method{names}{mergedCytoSignal}(x)

\method{names}{mergedCytoSignal}(x) <- value
}
\arguments{
\item{x}{A mergedCytoSignal object}

\item{value}{A character vector of up to the same length as x to rename
datasets. or NULL to reset to default names.}
}
\value{
For \code{names}, a character vector of dataset names.

For \code{names<-}, a mergedCytoSignal object with dataset names updated
in both metadata and the rownames of lrscore matrices
}
\description{
Allows getting and setting dataset names of a mergedCytoSignal object.
}
