% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{normCounts.CytoSignal}
\alias{normCounts.CytoSignal}
\title{Sub function for normCounts, input a Cytosignal object}
\usage{
\method{normCounts}{CytoSignal}(object, method = c("default", "cpm"), slot.use = NULL)
}
\arguments{
\item{object}{A Cytosignal object}

\item{method}{The method to use for normalization}

\item{slot.use}{The slot to use for normalization}
}
\value{
A Cytosignal object
}
\description{
Sub function for normCounts, input a Cytosignal object
}
