% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{normCounts.dgCMatrix}
\alias{normCounts.dgCMatrix}
\title{Sub function for normCounts, input a sparse matrix}
\usage{
\method{normCounts}{dgCMatrix}(mat, method = c("default", "cpm"))
}
\arguments{
\item{mat}{A sparse matrix}

\item{method}{The method to use for normalization}
}
\value{
A sparse matrix
}
\description{
Sub function for normCounts, input a sparse matrix
}
