% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{normCounts.list}
\alias{normCounts.list}
\title{Sub function for normCounts, input a sparse matrix}
\usage{
\method{normCounts}{list}(mat.list, method = c("default", "cpm"))
}
\arguments{
\item{mat.list}{list of (mat, scale.fac)}

\item{method}{The method to use for normalization}
}
\value{
A sparse matrix
}
\description{
Sub function for normCounts, input a sparse matrix
}
