% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{permuteImpLR}
\alias{permuteImpLR}
\title{Permute Imputation Results of specific imputation method}
\usage{
permuteImpLR(object, nn.type = NULL, times = 10)
}
\arguments{
\item{object}{A Cytosignal object}

\item{nn.type}{The imputation method to use}

\item{times}{Number of permutations}
}
\value{
A Cytosignal object
}
\description{
This function is a follow-up function of imputeNiche, which permutes the default or user-defined
imputation method and stored the results in the ImpData object.
}
