% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.r
\name{permuteListImpLR}
\alias{permuteListImpLR}
\title{Permute Imputation Results of specific imputation method}
\usage{
permuteListImpLR(object, nn.type = NULL, perm.size = 1e+05)
}
\arguments{
\item{object}{A Cytosignal object}

\item{nn.type}{The imputation method to use}

\item{perm.size}{Size of the permutation test}
}
\value{
A Cytosignal object
}
\description{
This function is a follow-up function of imputeNiche, which permutes the default or user-defined
imputation method and stored the results in the ImpData object.
}
