% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRscores.r
\name{permuteLR}
\alias{permuteLR}
\title{Permute Imputation Results of specific imputation method}
\usage{
permuteLR(object, slot.use = NULL, perm.size = 1e+05)
}
\arguments{
\item{object}{A Cytosignal object}

\item{slot.use}{The \code{lrscore} slot to use}

\item{perm.size}{Size of the permutation test}
}
\value{
A Cytosignal object
}
\description{
This function is a conveninent wrapper of \code{permuteLR.sparse}, which permutes the imputation methods
from which a given LR score is calculated. Note that all rounds of permutation will use the same
shuffle and sample index.
}
