% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRscores.r
\name{permuteLR.sparse}
\alias{permuteLR.sparse}
\title{Permute imputation data on corresponded imputation methods, using the same shuffle index}
\usage{
permuteLR.sparse(
  object,
  nn.type = NULL,
  perm.index = NULL,
  sample.index = NULL
)
}
\arguments{
\item{object}{A Cytosignal object}

\item{nn.type}{The imputation method to use}

\item{perm.size}{Size of the permutation test}
}
\value{
A dgCMatrix object
}
\description{
This function permutes the default or user-defined imputation method and stored the results in the
default or user-defined ImpData object. A index vector is needed to permute the data.
}
