% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{permuteScoreLR}
\alias{permuteScoreLR}
\title{Permute LR score for specific ligand-receptor imputation obj pairs}
\usage{
permuteScoreLR(object, slot.use = NULL)
}
\arguments{
\item{object}{A Cytosignal object}

\item{slot.use}{The imputation method to use}
}
\value{
A Cytosignal object
}
\description{
This function is a follow-up function of graphNicheLR. It computes the NUL LRscores
using the NULL imputation results and stores the results in the LR score object.
The null distribution of the LR scores can be used to test the significance of the LR scores.
}
