% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circosplot.R
\name{plotCircosNIntr}
\alias{plotCircosNIntr}
\title{Make Circos plot summarising number of interactions between clusters}
\usage{
plotCircosNIntr(
  object,
  lrscore.use = NULL,
  intr.use = NULL,
  colors.list = NULL,
  titles = NULL,
  title.size = 12,
  label.size = 8,
  circle.margin = c(0.01, 0.01, 0.01, 0.2)
)
}
\arguments{
\item{object}{A \code{\linkS4class{cytosignal}} object with significant
interactions inferred with \code{\link{inferSignif}}.}

\item{lrscore.use}{Which LR score(s) to use. Use the name specified with
\code{lrscore.name} when running \code{\link{inferScoreLR}}. Default
\code{NULL} uses all available smoothed scores.}

\item{intr.use}{Which interactions to use focus on. Default \code{NULL} uses
all significant interactions inferred for the selected \code{lrscore.use}.}

\item{colors.list}{A named list of colors for the clusters. Default
\code{NULL} uses built-in color palette.}

\item{title.size, label.size}{Font size of the title and cluster labels,
respectivey. Default \code{12} and \code{8}.}

\item{circle.margin}{A positive numeric vector of 4 values for the left,
right, bottom, top margin sizes of the main circos panel. Usually the top
(4th) value should be larger to accommodate the title. Default
\code{c(0.01, 0.01, 0.01, 0.2)}.}
}
\value{
NULL value returned. Circos plots are drawn.
}
\description{
This function counts all the edges from the neighborhood graph connecting
between cluster pairs, weighted by the number of significant
interactions detected in the receiver cells. This is then normalized by the
total number of possible edges between the cluster pairs as is all
interactions are significant.
Users can make multiple circos plots for different types of LR scores (e.g.
diffusion or contact dependent) at the same time, combined by grid.

This function behaves a bit differently when showing a single selected type
of interactions versus multiple types. When multiple plots are intended, by
either explicitly selecting multiple \code{lrscore.use} or by default when
multiple types of inference are available, the plots are automatically
arranged in a grid layout. In single-plot mode, the plot is filled into the
current graphics device. That means users can globally make layout settings
before calling and combine with other plots. See examples.
}
\examples{
\dontrun{
# Assuming `cs` is a cytosignal object after calling `inferSignif()`
# Simply show the count of diffusion-dependent interactions between cluster
# pairs.
plotCircosNIntr(cs, 'diffusion-Raw_smooth')

# Show all the count of all types of interactions. By default, an additional
# inference called "Raw-Raw" is also included for a contact-dependent
# interactions
plotCircosNIntr(cs)

# Show selected types of interactions only
plotCircosNIntr(cs, lrscore.use = c('diffusion-Raw_smooth',
                                   'diffusion-Env_smooth'))

# Customized layout for combining with other R graphics or re-arranging the
# plots for multiple types of interactions
par(mfrow = c(1,2))
plotCircosNIntr(cs, 'contact-Raw_smooth')
plotCircosNIntr(cs, 'diffusion-Raw_smooth')
}

}
