% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotCircosNIntr2}
\alias{plotCircosNIntr2}
\title{Make Circos plot summarising number of interactions between clusters}
\usage{
plotCircosNIntr2(
  object,
  type = NULL,
  splitType = FALSE,
  interaction = NULL,
  fdrThresh = 0.05,
  minExp = 100,
  minSpot = 100,
  colors = NULL,
  titles = NULL,
  titleTextSize = 12,
  labelTextSize = 8,
  circleMargin = c(0.01, 0.01, 0.01, 0.2)
)
}
\arguments{
\item{object}{A \code{\linkS4class{cytosignal}} object with significant
interactions inferred with \code{\link{inferSignif}}.}

\item{type}{Selecting the type of interaction to show. Default \code{NULL}
considers both options of \code{'contact'} and \code{'diffusion'}.}

\item{splitType}{Whether to split the interactions by type and show multiple
plots. Default \code{FALSE}. See details.}

\item{interaction}{Which interactions to use focus on. Default \code{NULL}
uses all significant interactions selected by \code{\link{getTopIntr}}.
Otherwise, users can provide a vector of interaction selection (name or
indices applicable to \code{intrDB(object)}), or a pre-filtered \code{csdb}
object (e.g. non-default \code{getTopIntr} result).}

\item{fdrThresh, minExp, minSpot}{Parameters passed to \code{\link{getTopIntr}}
for filtering significant interactions when \code{interaction = NULL}.
\code{fdrThresh} is additionally applied to claim significant interactions in
each spot.}

\item{colors}{A vector of colors for the clusters. If \code{NULL}, a built-in
color palette is used. Default \code{NULL}.}

\item{titles}{Customized title texts for the subplots. Default \code{NULL}
decides the titles automatically. When multiple plots are generated, the
lenght of \code{titles} must match the number of plots.}

\item{titleTextSize, label.size}{Font size of the title and cluster labels,
respectivey. Default \code{12} and \code{8}.}

\item{circleMargin}{A positive numeric vector of 4 values for the left,
right, bottom, top margin sizes of the main circos panel. Usually the top
(4th) value should be larger to accommodate the title. Default
\code{c(0.01, 0.01, 0.01, 0.2)}.}
}
\value{
NULL value returned. Circos plots are drawn.
}
\description{
This function counts all the edges from the neighborhood graph connecting
between cluster pairs, weighted by the number of significant
interactions detected in the receiver cells. This is then normalized by the
total number of possible edges between the cluster pairs as if all
interactions are significant.
}
\details{
\code{splitType} controls if all interactions are counted together in one
plot or counted separately and shown in multiple plots. When splitting by
type, the count is normalized by total number of interactions of that type
only.

When making a single plot, users can make use of R's graphic layout system
to combine with other plots or arrange the plot in a customized way. See
examples. When making multiple plot in one call, the plots are arranged by
default and the function ignores any global layout settings.
}
\examples{
\dontrun{
# Assuming `cs` is a cytosignal2 object after calling `inferLRScore()`
# Summarize the count of all types of interactions in one plot.
plotCircosNIntr2(cs)

# Show the count of only diffusion-dependent interactions between cluster
# pairs.
plotCircosNIntr2(cs, 'diffusion')

# Show the type of both types of interactions in separate panels
plotCircosNIntr2(cs, splitType = TRUE)

# Customized layout for combining with other R graphics or re-arranging the
# plots for multiple types of interactions
par(mfrow = c(1,2))
plotCircosNIntr2(cs, 'diffusion')
plot(1:10, 1:10)
}

}
