% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_edge.r
\name{plotEdge.CytoSignal}
\alias{plotEdge.CytoSignal}
\title{Plotting edge for a given interaction from a CytoSignal object}
\usage{
\method{plotEdge}{CytoSignal}(
  object,
  plot_dir,
  intr,
  type = c("sender", "receiver"),
  edge.size = 1000,
  slot.use = NULL,
  signif.use = NULL,
  colors.list = NULL,
  plot.fmt = "png",
  title = NULL,
  use.cex = 0.1,
  use.shape = 16,
  line.width = 0.01,
  use.phi = 30,
  use.theta = -17,
  z.scaler = 0.03,
  z.pt.interval = 1,
  pt.stroke = 0.2,
  u_width = 6,
  u_hgt = 5,
  set.res = 400,
  return.plot = T
)
}
\value{
Plot of edge
}
\description{
Plotting edge for a given interaction from a CytoSignal object
}
