% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_edge.r
\name{plotEdge.matrix_like}
\alias{plotEdge.matrix_like}
\title{Sub function for plotEdge}
\usage{
\method{plotEdge}{matrix_like}(
  cells.loc,
  type = NULL,
  edge.size = 2000,
  nn.graph.sig,
  receiver.idx,
  col.fac,
  res.list,
  intr,
  title,
  use.cex,
  use.shape,
  line.width,
  use.phi,
  use.theta,
  z.scaler,
  z.pt.interval,
  pt.stroke,
  u_width,
  u_hgt,
  set.res
)
}
\value{
a plist object
}
\description{
Sub function for plotEdge
}
