% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_edge.r
\name{plotEdge}
\alias{plotEdge}
\title{Plotting edge for a given interaction from a CytoSignal object}
\usage{
plotEdge(object, ...)
}
\arguments{
\item{object}{CytoSignal object}

\item{plot_dir}{Directory to save plots}

\item{intr}{Interaction to plot}

\item{type}{Type of plot, either "sender" or "receiver"}

\item{slot.use}{Slot to use for plotting}

\item{signif.use}{Significance level to use for plotting}

\item{colors.list}{List of colors to use for plotting}

\item{plot.fmt}{Format of plot, either "png" or "pdf"}

\item{title}{Title of plot}

\item{use.cex}{Size of points}

\item{use.shape}{Shape of points}

\item{line.width}{Width of lines}

\item{use.phi}{Angle of 3D plot}

\item{use.theta}{Angle of 3D plot}

\item{z.scaler}{Scaling factor for z-axis}

\item{z.pt.interval}{Interval of points on z-axis}

\item{pt.stroke}{Width of points}

\item{u_width}{Width of plot}

\item{u_hgt}{Height of plot}

\item{set.res}{Resolution of plot}

\item{return.plot}{Whether to return plot}
}
\value{
Plot of edge
}
\description{
Plotting edge for a given interaction from a CytoSignal object
}
