% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotEdge2}
\alias{plotEdge2}
\title{3D plot showing edges connecting sender and receiver spots for selected interactions}
\usage{
plotEdge2(
  object,
  interaction,
  colors = NULL,
  fdrThresh = 0.05,
  showTitle = TRUE,
  showSubTitle = TRUE,
  showLegend = TRUE,
  showBox = TRUE,
  edgeLocalMinSize = 10,
  edgeSampleRate = 0.05,
  theta = -17,
  phi = 30,
  boxHeight = 0.7,
  perspTrans = 5,
  dotSizeSignif = 0.3,
  dotAlphaSignif = 1,
  dotSizeOther = 0.16,
  dotAlphaOther = 0.3,
  edgeColor = "grey",
  edgeLinewidth = 0.2,
  titleTextSize = 12,
  subtitleTextSize = 10,
  legendTitleSize = 10,
  legendTextSize = 8,
  legendNCol = NULL,
  labelTextSize = 8,
  seed = NULL
)
}
\arguments{
\item{object}{A \code{\linkS4class{cytosignal2}} object.}

\item{interaction}{Which interactions to use. Users can provide a vector of
interaction selection (name or indices applicable to rows of \code{intrDB(object)}).
Or a subset of the database itself (e.g. \code{getTopIntr(object, topN = 2, type = 'cont')})}

\item{colors}{Color vector to use for categorical or logical variables.
Default \code{NULL} adopts: when categorical, built-in colors \code{csColors}
that allows maximum 51 categories; when logical, red TRUE and grey FALSE.}

\item{fdrThresh}{FDR threshold to claim that an interaction is significantly
observed in a spot. Default \code{0.05}.}

\item{showTitle, showSubTitle, showLegend}{Logical, whether to show title,
subtitle, and legend in the plot. Default \code{TRUE}.}

\item{showBox}{Logical, whether to show the box frame lines. Axis arrows and
labels are also hidden when set to \code{FALSE}. Default \code{TRUE}.}

\item{edgeLocalMinSize}{Minimum number of edges to sample from each local
spatial cluster of receiver spots. This balances the downsampling when
satellite interaction activities are present. Default \code{10}.}

\item{edgeSampleRate}{Proportion of edges to sample from major interaction
activity areas. Default \code{0.05}.}

\item{theta}{3D plot horizontal rotation angle, positive value rotate
clockwise when looking from top. Default \code{-17}.}

\item{phi}{3D plot vertical rotation angle, positive value rotate clockwise
when looking from left side, or say the side closer to users goes downward.
Default \code{30}.}

\item{boxHeight}{Height of the 3D box, a value between 0 and 1. Default
\code{0.7}.}

\item{perspTrans}{A value which can be used to vary the strength of the
perspective transformation. Values greater than 1 will lessen the perspective
effect and values less and 1 will exaggerate it. Default \code{5}.}

\item{dotSizeSignif, dotAlphaSignif}{Dot size and transparency for spots
to be highlighted. Default \code{0.3} and \code{1} (non-transparent).}

\item{dotSizeOther, dotAlphaOther}{Dot size and transparency for other spots.
Default \code{0.16} and \code{0.3} (mostly transparent).}

\item{edgeColor}{Color of the edges. Default \code{'grey'}.}

\item{edgeLinewidth}{Line width of the edges. Default \code{0.2}.}

\item{titleTextSize, subtitleTextSize, legendTextSize}{Font sizes for the plot
title, subtitle, and legend text. Default \code{12}, \code{10}, and \code{8},
respectively.}

\item{legendTitleSize}{Font size for the legend title. Default \code{10}.}

\item{labelTextSize}{Font size for the z-axis label. Default \code{8}.}

\item{seed}{Random seed for edge sampling for reproducibility. Default
\code{NULL} does not set seed. Setting a seed does not interfere with R's
global random state.}
}
\value{
A \code{plot3D} plist object when one interaction is provided; A list
of \code{plot3D} plist objects when multiple interactions are provided. A
plist object can be printed in console and shown in viewer, or saved as RDS
file for later use.
}
\description{
This function selects the receiver spots with significant LRscores for
the selected interactions, and connects them to corresponding neighbor sender
spots with positive ligand expression.

A spatial coordinate panel colored by cluster is shown at bottom and top
plains of the 3D plot, with solid dots representing selected sender/receiver
spots, and translucent dots representing other spots. The edges connecting
the sender and receiver spots are shown as segments between the two plains.
The edges are downsampled for clearer visualization.
}
