% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotEdgeDot}
\alias{plotEdgeDot}
\title{Dot plot showing cluster-wise counts of edges between spots for each interaction}
\usage{
plotEdgeDot(
  object,
  interaction,
  senders = NULL,
  receivers = NULL,
  sumIntrs = FALSE,
  fdrThresh = 0.05
)
}
\arguments{
\item{object}{A \code{\linkS4class{cytosignal2}} object.}

\item{interaction}{Which interactions to use. Users can provide a vector of
interaction selection (name or indices applicable to rows of \code{intrDB(object)}).
Or a subset of the database itself (e.g. \code{getTopIntr(object, topN = 2, type = 'cont')})}

\item{senders}{Character vector of sender cluster names to only shown in the
plot. Default \code{NULL} shows all clusters that send valid signal.}

\item{receivers}{Character vector of receiver cluster names to only shown in
the plot. Default \code{NULL} shows all clusters that receive valid signal.}

\item{sumIntrs}{Logical, whether to sum the edges from all selected
interactions into one plot. Default \code{FALSE} generates separate plots for
each interaction.}

\item{fdrThresh}{FDR threshold to claim that an interaction is significantly
observed in a spot. Default \code{0.05}.}
}
\value{
A \code{ggplot} object when \code{sumIntrs = FALSE} and only one
interaction is selected; A list of \code{ggplot} objects when
\code{sumIntrs = FALSE} and multiple interactions are selected; A single
\code{ggplot} object when \code{sumIntrs = TRUE}.
}
\description{
This function collects all valid sender and receiver spot pairs for each
interaction, and summarizes the counts of such pairs (edges) by sender
cluster and receiver cluster. This function further calculates the total
expected possible edges between clusters based on the neighbor graph
corresponding to the type of each interaction. A proportion is then derived
by dividing the observed edge counts by the expected possible counts.
}
