% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intrDEG.R
\name{plotIntrDEGHeatmap}
\alias{plotIntrDEGHeatmap}
\title{Heatmap visualization of interaction-associated DEG expression}
\usage{
plotIntrDEGHeatmap(
  object,
  intrDEG,
  interaction = NULL,
  topN = 20,
  sampleSize = 5000,
  clusterLabelSize = 8,
  geneNameSize = 8,
  rowTitleSize = 10,
  annNameSize = 8,
  titleSize = 12,
  legendTitleSize = 10,
  legendTextSize = 8
)
}
\arguments{
\item{object}{A \code{\linkS4class{cytosignal2}} object.}

\item{intrDEG}{A \code{csIntrDEG} object returned by
\code{\link{findIntrDEG}} or \code{\link{refineIntrDEG}}.}

\item{interaction}{Which interactions to use. Users can provide a vector of
interaction selection (name or indices applicable to rows of \code{intrDB(object)}).
Or a subset of the database itself (e.g. \code{getTopIntr(object, topN = 2, type = 'cont')})}

\item{topN}{Maximum number of top DEGs to show per interaction. Default
\code{20}.}

\item{sampleSize}{Number of spots to randomly sample for visualization.
Default \code{5000}.}

\item{clusterLabelSize}{Font size for cluster labels at the bottom. Default
\code{8}.}

\item{geneNameSize}{Font size for gene names on the right. Default \code{8}.}

\item{rowTitleSize}{Font size for row titles on the left, such as "ligand",
"receptor" and "DEG". Default \code{10}.}

\item{annNameSize}{Font size for annotation names on the right of the top
annotation bar. Default \code{8}.}

\item{titleSize}{Font size for the heatmap title. Default \code{12}.}

\item{legendTitleSize}{Font size for the legend title. Default \code{10}.}

\item{legendTextSize}{Font size for the legend text. Default \code{8}.}
}
\value{
A \code{HeatmapList} object powered by package "ComplexHeatmap" if
only one interaction is plotted. A list of \code{HeatmapList} objects if
multiple interactions are plotted.
}
\description{
This function makes heatmaps to visualize the expression of
interaction-associated DEGs detected by \code{\link{findIntrDEG}} and/or
\code{\link{refineIntrDEG}}. For each interaction, a heatmap is made to show
the expression of the ligand and receptor genes, and the DEGs. The spots
are grouped by their cluster labels and whether they receive significant
interaction signal.
}
\details{
The top \code{topN} DEGs are directly sliced from
\code{intrDEG$$geneSignif[["interaction_name"]]}, as the table is pre-sorted.
When using a \code{csIntrDEG} object returned by \code{\link{refineIntrDEG}},
the clusters that show association with the interaction will be highlighted
in black text at the bottom of the heatmap, while other clusters are colored
in grey.
}
\seealso{
\code{\link{findIntrDEG}}, \code{\link{refineIntrDEG}}
}
