% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotIntrSummary}
\alias{plotIntrSummary}
\title{Combined plots of all information for selected interactions}
\usage{
plotIntrSummary(
  object,
  interaction,
  fdrThresh = 0.05,
  edge = FALSE,
  velo = FALSE,
  legendNCol = NULL,
  legendNRow = NULL,
  colors = NULL,
  legendDotSize = 4,
  dotSize = 0.5,
  dotAlpha = 1,
  paletteOption = "C",
  paletteDirection = -1,
  zeroAsNA = TRUE,
  naColor = "grey80",
  titleTextSize = 12,
  subtitleTextSize = 10,
  legendTextSize = 8
)
}
\arguments{
\item{object}{A \code{\linkS4class{cytosignal2}} object.}

\item{interaction}{Which interactions to use. Users can provide a vector of
interaction selection (name or indices applicable to rows of \code{intrDB(object)}).
Or a subset of the database itself (e.g. \code{getTopIntr(object, topN = 2, type = 'cont')})}

\item{fdrThresh}{FDR threshold to claim that an interaction is significantly
observed in a spot. Default \code{0.05}.}

\item{edge, velo}{Not implemented yet}

\item{legendNCol, legendNRow}{Controls the layout of cluster legend. When only
two genes are to be shown (one ligand and one receptor), only
\code{legendNCol} is used to set the number of columns in the legend. When
more than two genes are to be shown (complex ligand/receptor), only
\code{legendNRow} is used to set the number of rows in the legend.}

\item{colors}{Color vector to use for categorical or logical variables.
Default \code{NULL} adopts: when categorical, built-in colors \code{csColors}
that allows maximum 51 categories; when logical, red TRUE and grey FALSE.}

\item{legendDotSize}{Dot size in the legend for categorical or logical
variables. Default \code{4}.}

\item{dotSize}{Size of the dots. Default \code{0.5}.}

\item{dotAlpha}{Transparency of the dots in the scatter plot. Default
\code{1} for no transparency. 0 for fully transparent.}

\item{paletteOption}{Color palette option for continuous variables. One of
letter from A to H, or exact name from \code{magma}, \code{inferno},
\code{plasma}, \code{viridis}, \code{cividis}, \code{rocket}, \code{mako},
or \code{turbo}.}

\item{paletteDirection}{Direction of the color palette for continuous
variables. Default \code{-1} for dark-solid colors representing high values.
Use \code{1} for light colors representing high values.}

\item{zeroAsNA}{Logical, whether to treat zero values as NA for continuous
variables. Works together with argument \code{naColor} to distinguish
zero value from low positive values. Default \code{TRUE}.}

\item{naColor}{Color to use for NA values, and NA values converted from
zeros. Default \code{'grey80'}.}

\item{titleTextSize, subtitleTextSize, legendTextSize}{Font sizes for the plot
title, subtitle, and legend text. Default \code{12}, \code{10}, and \code{8},
respectively.}
}
\value{
A \code{\link[patchwork]{patchwork}} object when one interaction is
provided; A list of \code{\link[patchwork]{patchwork}} objects when multiple
interactions are provided.
}
\description{
This function generates combined plots showing spatial expression of
ligand and receptor genes, spatial LR score, and cluster annotation for
selected interactions.
}
