% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nebula.R
\name{plotNebulaAll}
\alias{plotNebulaAll}
\title{Visualize NEBULA analysis result with dot plot that involves all covariates}
\usage{
plotNebulaAll(
  object,
  intr.type = c("diff", "cont"),
  fdrThresh = 0.05,
  logfcThresh = 1,
  title = NULL,
  titleSize = 16,
  xTextSize = 10,
  xTextAngle = 40,
  yTextSize = 10,
  yTitleSize = 12,
  legendTextSize = 12,
  legendTitleSize = 12,
  stripTitleSize = 12,
  ylim = NULL
)
}
\arguments{
\item{object}{A \linkS4class{mergedCytoSignal} object, with
\code{\link{runNEBULA}()} run in advance.}

\item{intr.type}{A character string specifying the type of interaction to
plot. Choose from \code{"diff"} (default) for diffusion-dependent or
\code{"cont"} for contact-dependent interactions.}

\item{fdrThresh}{A numeric value to set the FDR threshold for significance.
FDR value higher than this threshold will be considered as not significant.
Default \code{0.05}.}

\item{logfcThresh}{A numeric value to set the threshold on the absolute value
of logFC. Absoluate logFC value lower than this threshold will be considered
as not significant. Default \code{1}.}

\item{title}{A character string to set the title of the plot. Default
\code{NULL} does not add a title.}

\item{titleSize}{A numeric value to set the size of the title of the plot.
Default \code{16}.}

\item{xTextSize}{A numeric value to set the size of the text on x-axis.
Default \code{10}.}

\item{xTextAngle}{A numeric value to set the angle (in degree) to rotate the
x-axis text labels. Default \code{40}.}

\item{yTextSize}{A numeric value to set the size of the text on y-axis.
Default \code{10}.}

\item{yTitleSize}{A numeric value to set the size of the title on y-axis.
Default \code{12}.}

\item{legendTextSize}{A numeric value to set the size of the text in the
legend. Default \code{12}.}

\item{legendTitleSize}{A numeric value to set the size of the title in the
legend. Default \code{12}.}

\item{stripTitleSize}{A numeric value to set the size of the title in the
strip on top of each grid for the covariate groups. Default \code{12}.}

\item{ylim}{Numeric vector with 2 numbers to set the limits of y-axis.
Default \code{NULL} shows the full figure.}
}
\value{
A ggplot figure (\code{gg}, \code{ggplot} object).
}
\description{
This visualization method uses the result for either diffusion-dependent or
contact-dependent interactions from NEBULA analysis to generate a dot plot.
Each dot represents an interaction. The x axis is grouped by covariate levels
and jittered to avoid overlap. Therefore, each interaction appears in all the
covariate groups. The y axis is the logFC value of the interaction. The size
of the dots represents the significance (\eqn{-log_{10}FDR}) of an
interaction evaluated on a covariate level. Transparency of the dots shows
overall significance combining the FDR threshold and logFC threshold.
}
