% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nebula.R
\name{plotNebulaVolcano}
\alias{plotNebulaVolcano}
\title{Make volcano plot for NEBULA analysis result}
\usage{
plotNebulaVolcano(
  object,
  covariate,
  intr.type = c("diff", "cont"),
  highlight = NULL,
  topN = NULL,
  fdrThresh = 0.05,
  logfcThresh = 1,
  interactive = TRUE,
  dotSize = 1.3,
  dotAlpha = 0.8,
  xTextSize = 10,
  xTitleSize = 12,
  yTextSize = 10,
  yTitleSize = 12,
  legendTextSize = 12,
  legendTitleSize = 12,
  labelTextSize = 4,
  titleSize = 16
)
}
\arguments{
\item{object}{A \linkS4class{mergedCytoSignal} object, with
\code{\link{runNEBULA}()} run in advance.}

\item{covariate}{A character string of the covariate to plot. Must be
available in \code{showCov(object)}.}

\item{intr.type}{A character string specifying the type of interaction to
plot. Choose from \code{"diff"} (default) for diffusion-dependent or
\code{"cont"} for contact-dependent interactions.}

\item{highlight}{A character vector of interaction names, a numeric or
logical vector of indices to select specific interactions to highlight; or
a single \code{TRUE} to highlight all significant interactions. See
\code{showIntr(object, intr.type)} for available options. Default \code{NULL}.}

\item{topN}{An integer to select top N interactions to highlight ranked by
significance. Rank by FDR first and abs(logFC) then. Default \code{NULL}.}

\item{fdrThresh}{A numeric value to set the FDR threshold for significance.
FDR value higher than this threshold will be considered as not significant.
Default \code{0.05}.}

\item{logfcThresh}{A numeric value to set the threshold on the absolute value
if logFC. Absoluate logFC value lower than this threshold will be considered
as not significant. Default \code{1}.}

\item{interactive}{Logical, whether to generate an interactive plotly figure.
Default \code{TRUE}. Generate a static ggplot figure if \code{FALSE}.}

\item{dotSize}{A numeric value to set the size of the dots in the plot.
Default \code{1.3}.}

\item{dotAlpha}{A numeric value to set the transparency of the dots in the
plot. Default \code{0.8}.}

\item{xTextSize}{A numeric value to set the size of the text on x-axis.
Default \code{10}.}

\item{xTitleSize}{A numeric value to set the size of the title on x-axis.
Default \code{12}.}

\item{yTextSize}{A numeric value to set the size of the text on y-axis.
Default \code{10}.}

\item{yTitleSize}{A numeric value to set the size of the title on y-axis.
Default \code{12}.}

\item{legendTextSize}{A numeric value to set the size of the text in the
legend. Default \code{12}.}

\item{legendTitleSize}{A numeric value to set the size of the title in the
legend. Default \code{12}.}

\item{labelTextSize}{A numeric value to set the size of the text on the
highlight labels. Default \code{4}.}

\item{titleSize}{A numeric value to set the size of the main title of the
plot. Default \code{16}.}
}
\value{
\itemize{
\item A plotly figure (\code{plotly}, \code{htmlwidget} object) if \code{interactive = TRUE}.
\item A ggplot figure (\code{gg}, \code{ggplot} object) if \code{interactive = FALSE}.
}
}
\description{
This function allows drawing a volcano plot for the Nebula analysis on a
single covariate level, for either diffusion-dependent or contact-dependent
interactions. Either static ggplot figure or interactive plotly figure can
be generated. Interactions of interests can be labeled with: 1. choosing
specific interactions by name or index. 2. choosing top N interactions ranked
by significance (FDR and abs(logFC)). A non-NULL \code{highlight} argument is
considered before \code{topN}. When both are not specified, highlight all
significant interactions.
}
