% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor.R
\name{plotNeighborDiff}
\alias{plotNeighborDiff}
\alias{plotNeighborCont}
\title{Visualize neighbor weights of a spot}
\usage{
plotNeighborDiff(
  object,
  index = NULL,
  dotSize = 0.5,
  dotAlpha = 1,
  paletteOption = "D",
  paletteDirection = -1,
  zeroAsNA = TRUE,
  naColor = "grey70"
)

plotNeighborCont(
  object,
  index = NULL,
  dotSize = 0.5,
  dotAlpha = 1,
  paletteOption = "D",
  paletteDirection = -1,
  zeroAsNA = TRUE,
  naColor = "grey70"
)
}
\arguments{
\item{object}{A \code{cytosignal2} object. With \code{\link{findNeighbor}} called.}

\item{index}{Index of the spot to visualize. Default \code{NULL} selects
the spot with the most neighbors. Can be an integer index, character ID, or
logical selector vector.}

\item{dotSize}{Size of the dots. Default \code{0.5}.}

\item{dotAlpha}{Transparency of the dots in the scatter plot. Default
\code{1} for no transparency. 0 for fully transparent.}

\item{paletteOption}{Color palette option for continuous variables. One of
letter from A to H, or exact name from \code{magma}, \code{inferno},
\code{plasma}, \code{viridis}, \code{cividis}, \code{rocket}, \code{mako},
or \code{turbo}.}

\item{paletteDirection}{Direction of the color palette for continuous
variables. Default \code{-1} for dark-solid colors representing high values.
Use \code{1} for light colors representing high values.}

\item{zeroAsNA}{Logical, whether to treat zero values as NA for continuous
variables. Works together with argument \code{naColor} to distinguish
zero value from low positive values. Default \code{TRUE}.}

\item{naColor}{Color to use for NA values, and NA values converted from
zeros. Default \code{'grey80'}.}
}
\value{
A ggplot2 object showing the spatial coordinates colored by the neighbor
weights of the selected spot.
}
\description{
These functions make spatial coordinate plots colored by the weights of a
specified spot in the selected type of neighborhood graph. By default,
the spot with the most neighbors is selected.
}
\seealso{
\code{\link[=findNeighbor]{findNeighbor()}}
}
