% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_basic.R
\name{plotTotalCountQC}
\alias{plotTotalCountQC}
\alias{plotGeneDetectedQC}
\title{Quality control plotting functions}
\usage{
plotTotalCountQC(object, base_size = 11, statsTextSize = base_size/3)

plotGeneDetectedQC(object, base_size = 11, statsTextSize = base_size/3)
}
\arguments{
\item{object}{A \code{\linkS4class{cytosignal2}} object.}

\item{base_size}{Base font size. Default \code{11}.}

\item{statsTextSize}{Font size of the labeled statistics. Default is
\code{base_size/3}.}
}
\value{
A ggplot object showing the QC metric.
}
\description{
These are the functions for plotting quality control (QC) metrics via violin
plots, with boxplots overlaid and key statistics labeled.
}
