% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_gene.R
\name{plotRefinedScore}
\alias{plotRefinedScore}
\title{Plot the refined score of each interaction after regression model refinement}
\usage{
plotRefinedScore(
  object,
  intrDEGRes,
  intr = NULL,
  pt.size = 0.5,
  pt.stroke = 0.1
)
}
\arguments{
\item{object}{A \code{Cytosignal} object with \code{\link{inferIntrScore}}
already run.}

\item{intrDEGRes}{The direct output object of \code{\link{inferIntrDEG}}.}

\item{intr}{A vector of unique interaction IDs that are available in
\code{intrDEGRes}, or numerical index within its range. Default \code{NULL}
use all the results.}

\item{pt.size}{Size of the points in the plot. Default \code{0.5}.}

\item{pt.stroke}{Stroke size of the points in the plot. Default \code{0.1}.}
}
\value{
List of ggplot objects, each shows the refined LR score of each
selected interaction.
}
\description{
Plot the refined score of each interaction after regression model refinement
}
