% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REVIGO.R
\name{plotREVIGO}
\alias{plotREVIGO}
\title{Create scatter plot from REVIGO result}
\usage{
plotREVIGO(
  rvg,
  labelBy = c("Name", "TermID"),
  labelIdx = NULL,
  labelSize = 4,
  max.overlaps = 5
)
}
\arguments{
\item{rvg}{A data.frame from \code{\link{revigo}}. Must contain columns:
\code{PC_0}, \code{PC_1}, \code{LogSize}, \code{Frequency}, and \code{Name}.}

\item{labelBy}{From which field of the revigo result to retrieve label text.
Choose from \code{"Name"} or \code{"TermID"}. Default \code{"Name"}.}

\item{labelIdx}{Index matching to rows of \code{rvg} to indicate whether a
label has to be shown for the bubble. Default \code{NULL} automatically
shows tidy view with ggrepel.}

\item{labelSize}{Size of text label for marked dots. Default \code{4}.}

\item{max.overlaps}{\code{?ggrepel::geom_text_repel} parameter. Default
\code{5}.}
}
\value{
A ggplot object
}
\description{
Create scatter plot from REVIGO result
}
\examples{
rvg <- revigo(c("GO:0003002", "GO:0007389", "GO:0007423", "GO:0009653"),
              speciesTaxon = 10090)
plotREVIGO(rvg)
}
