% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_basic.R
\name{plotScale}
\alias{plotScale}
\title{Check if parameters look right with scale bar on spatial coordinates}
\usage{
plotScale(
  object,
  position = c("bl", "br", "tl", "tr"),
  direction = c("horizontal", "vertical"),
  labelSize = 5
)
}
\arguments{
\item{object}{A \code{cytosignal2} object.}

\item{position}{Position of the scale bar. One of \code{'bl'} (bottom left),
\code{'br'} (bottom right), \code{'tl'} (top left), or \code{'tr'} (top
right). Default \code{'bl'}.}

\item{direction}{Direction of the scale bar. Either \code{'horizontal'} or
\code{'vertical'}. Default \code{'horizontal'}.}

\item{labelSize}{Size of the labeled text. Default \code{5}.}
}
\value{
A ggplot2 object showing the spatial coordinates with a scale bar.
Error is thrown if the parameters have not been set.
}
\description{
Make a spatial coordinate plot with a scale bar added. The length of the
scale bar represents the radius of the Epsilon ball, which is used later for
modeling the neighborhood for diffusion-dependent LR interactions. Users can
determine if the scale bar looks appropriate based on the experimental setup
of the sliced tissue.
}
