% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_edge.r
\name{plotSigCluster}
\alias{plotSigCluster}
\title{Plot edges by each cluster as sender or receiver cells}
\usage{
plotSigCluster(
  object,
  plot_dir,
  cluster.list = NULL,
  intr.num = 10,
  type = c("sender", "receiver"),
  slot.use = NULL,
  signif.use = NULL,
  colors.list = NULL,
  plot.fmt = "png",
  edge.size = 2000,
  all.in.one = T,
  plot.all.sig = F,
  use.cex = 0.1,
  use.shape = 16,
  line.width = 0.02,
  use.phi = 30,
  use.theta = -17,
  z.scaler = 0.03,
  z.pt.interval = 1,
  pt.stroke = 0.2,
  u_width = 6,
  u_hgt = 6,
  set.res = 400,
  return.plot = F
)
}
\arguments{
\item{object}{A \code{CytoSignal} object}

\item{plot_dir}{Directory to save plots}

\item{cluster.list}{A list of clusters to plot}

\item{intr.num}{Number of interactions to plot}

\item{type}{Either "sender" or "receiver"}

\item{slot.use}{Slot to use for plotting}

\item{signif.use}{Significance threshold to use for plotting}

\item{colors.list}{A list of colors to use for each cluster}

\item{plot.fmt}{Plot format}

\item{edge.size}{Number of edges to plot}

\item{all.in.one}{Plot all clusters in one plot}

\item{plot.all.sig}{Plot all significant edges}

\item{use.shape}{Shape of points}

\item{line.width}{Width of lines}

\item{use.phi}{Angle of view}

\item{use.theta}{Angle of view}

\item{z.scaler}{Scale of z-axis}

\item{z.pt.interval}{Interval of z-axis}

\item{pt.stroke}{Stroke of points}

\item{u_width}{Width of plot}

\item{u_hgt}{Height of plot}

\item{set.res}{Resolution of plot}

\item{return.plot}{Return plot object}

\item{pt.size}{Size of points}

\item{...}{Other arguments}
}
\value{
A list of plots
}
\description{
By default ranked by the number of significant cells in each cluster
}
