% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{plotSignif}
\alias{plotSignif}
\title{Plot significant interactions ranked by the user-specified metric}
\usage{
plotSignif(
  object,
  num.plot = NULL,
  res_dir,
  plot.details = T,
  slot.use = NULL,
  signif.use = NULL,
  plot.clusters = T,
  plot.velo = F,
  colors.list = NULL,
  pt.size = 0.5,
  pt.stroke = 0.2,
  u_width = 6,
  u_hgt = 5,
  set.res = 200,
  return.plot = F
)
}
\arguments{
\item{object}{A cytosignal object}

\item{num.plot}{Number of interactions to plot}

\item{res_dir}{Directory to save the plots}

\item{plot.details}{Whether to plot NULL imputed values and scores}

\item{slot.use}{The LRscore slot to use for plotting}

\item{signif.use}{The metric used to rank the interactions, by default "result.hq.pear"}

\item{plot.clusters}{Whether to plot the clusters}

\item{plot.velo}{Whether to plot the velocity}

\item{colors.list}{A list of colors to use for plotting}

\item{pt.size}{Size of the points}

\item{pt.stroke}{Stroke of the points}

\item{u_width}{Width of the plot}

\item{u_hgt}{Height of the plot}

\item{set.res}{Resolution of the plot}

\item{return.plot}{Whether to return the plot}
}
\value{
A plot if return.plot is TRUE. Otherwise, plots are saved to the specified directory.
}
\description{
Plot significant interactions ranked by the user-specified metric
}
