% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{plotSignif2}
\alias{plotSignif2}
\title{Plot significant interactions ranked by the user-specified metric}
\usage{
plotSignif2(
  object,
  intr,
  intr.rank,
  edge = FALSE,
  velo = FALSE,
  slot.use = NULL,
  signif.use = NULL,
  colors.list = NULL,
  pt.size = 0.1,
  pt.stroke = 0.2,
  return.plot = FALSE,
  plot_dir = "csSignifPlot/",
  plot.fmt = c("png", "pdf", "svg"),
  resolution = 200,
  verbose = FALSE
)
}
\arguments{
\item{object}{A cytosignal object}

\item{intr, intr.rank}{Specify interactions to be plotted. \code{intr} for
names \code{intr.rank} for numeric index.}

\item{edge, velo}{Logical, whether to plot edge or velocity, respectively.}

\item{slot.use}{The LRscore slot to use for plotting}

\item{signif.use}{The metric used to rank the interactions, by default "result.hq.pear"}

\item{colors.list}{A list of colors to use for plotting}

\item{pt.size}{Size of the points}

\item{pt.stroke}{Stroke of the points}

\item{return.plot}{Whether to return the plot}

\item{plot_dir}{Directory to save the plots}

\item{plot.fmt}{Format of output file. "png", "pdf", or "svg".}

\item{resolution}{Resolution of the output figure.}

\item{verbose}{Logical, whether to show progress.}
}
\value{
A plot if return.plot is TRUE. Otherwise, plots are saved to the specified directory.
}
\description{
Plot significant interactions ranked by the user-specified metric
}
