% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_basic.R
\name{plotSpatial}
\alias{plotSpatial}
\alias{plotSpatialLRScore}
\alias{plotSpatialGene}
\alias{plotSpatialMetadata}
\title{Spatial coordinate plotting functions}
\usage{
plotSpatial(object, metadataCol = NULL, gene = NULL, interaction = NULL, ...)

plotSpatialLRScore(
  object,
  interaction,
  dotSize = 0.5,
  dotAlpha = 1,
  paletteOption = "C",
  paletteDirection = -1,
  zeroAsNA = TRUE,
  naColor = "grey80",
  titleTextSize = 12,
  subtitleTextSize = 10,
  legendTextSize = 8
)

plotSpatialGene(
  object,
  gene,
  dotSize = 0.5,
  dotAlpha = 1,
  paletteOption = "C",
  paletteDirection = -1,
  zeroAsNA = TRUE,
  naColor = "grey80",
  titleTextSize = 12,
  subtitleTextSize = 10,
  legendTextSize = 8
)

plotSpatialMetadata(
  object,
  metadataCol = NULL,
  dotSize = 0.5,
  dotAlpha = 1,
  colors = NULL,
  legendDotSize = 4,
  legendNCol = NULL,
  legendNRow = NULL,
  paletteOption = "D",
  paletteDirection = -1,
  zeroAsNA = TRUE,
  naColor = "grey80",
  titleTextSize = 12,
  subtitleTextSize = 10,
  legendTextSize = 8
)
}
\arguments{
\item{object}{A \code{\linkS4class{cytosignal2}} object.}

\item{metadataCol}{Metadata column name(s) as in the column names of
\code{object@metadata}.}

\item{gene}{Gene name(s) as in the row names of \code{object@rawData}.}

\item{interaction}{Interaction name(s) as in the column names of
\code{object@LRScore}.}

\item{...}{Graphic parameters passed down to specific functions in this same
family.}

\item{dotSize}{Size of the dots. Default \code{0.5}.}

\item{dotAlpha}{Transparency of the dots in the scatter plot. Default
\code{1} for no transparency. 0 for fully transparent.}

\item{paletteOption}{Color palette option for continuous variables. One of
letter from A to H, or exact name from \code{magma}, \code{inferno},
\code{plasma}, \code{viridis}, \code{cividis}, \code{rocket}, \code{mako},
or \code{turbo}.}

\item{paletteDirection}{Direction of the color palette for continuous
variables. Default \code{-1} for dark-solid colors representing high values.
Use \code{1} for light colors representing high values.}

\item{zeroAsNA}{Logical, whether to treat zero values as NA for continuous
variables. Works together with argument \code{naColor} to distinguish
zero value from low positive values. Default \code{TRUE}.}

\item{naColor}{Color to use for NA values, and NA values converted from
zeros. Default \code{'grey80'}.}

\item{titleTextSize, subtitleTextSize, legendTextSize}{Font sizes for the plot
title, subtitle, and legend text. Default \code{12}, \code{10}, and \code{8},
respectively.}

\item{colors}{Color vector to use for categorical or logical variables.
Default \code{NULL} adopts: when categorical, built-in colors \code{csColors}
that allows maximum 51 categories; when logical, red TRUE and grey FALSE.}

\item{legendDotSize}{Dot size in the legend for categorical or logical
variables. Default \code{4}.}

\item{legendNCol, legendNRow}{Number of columns and rows in the legend for
categorical or logical variables. Default \code{NULL} tries to keep at most
15 rows.}
}
\value{
If only one value is given to plot, a ggplot object is returned. If
multiple values are given, a patchwork object combining all plots is
returned. A patchwork wrap can be directly printed for combined view of all
plots. Users can also extract individual plots from the patchwork wrap by
e.g. \code{plotlist[[1]]}.
}
\description{
These are the functions for plotting any potential values stored in a
\code{\linkS4class{cytosignal2}} object that can be applied on to the spatial
coordinates.
}
