% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_velo.R
\name{plotVelo}
\alias{plotVelo}
\title{Plot 3D LR-velo ranked by the user-specified metric}
\usage{
plotVelo(
  object,
  intr,
  return.plot = TRUE,
  plot_dir = "csVeloPlot/",
  filename = NULL,
  plot.fmt = c("png", "pdf"),
  slot.use = NULL,
  signif.use = NULL,
  plot.clusters = TRUE,
  colors.list = NULL,
  z.scaler = 0.03,
  title = NULL,
  pt.size = 0.1,
  use.shape = 16,
  use_xbins = 15,
  use_ybins = 15,
  arrow.line.width = 0.6,
  arrow.width = 0.06,
  pt.stroke = 0.2,
  use.phi = 30,
  use.theta = -17,
  width = 6,
  height = 6,
  set.res = 300,
  verbose = TRUE
)
}
\arguments{
\item{object}{A cytosignal object}

\item{return.plot}{Whether to return the plot}

\item{slot.use}{The LRscore slot to use for plotting}

\item{signif.use}{The metric used to rank the interactions, by default "result.hq.pear"}

\item{plot.clusters}{Whether to plot the clusters}

\item{colors.list}{A list of colors to use for plotting}

\item{z.scaler}{Scaling factor for the z-axis}

\item{pt.size}{Point size}

\item{use.shape}{Point shape}

\item{use_xbins}{Number of bins for the x-axis}

\item{use_ybins}{Number of bins for the y-axis}

\item{arrow.line.width}{Width of the arrow line}

\item{arrow.width}{Width of the arrow}

\item{pt.stroke}{Stroke of the points}

\item{use.phi}{Set view angel: phi the colatitude}

\item{use.theta}{Set view angel: theta gives the azimuthal direction}

\item{width}{Width of the plot}

\item{height}{Height of the plot}

\item{set.res}{Resolution of the plot}

\item{num.plot}{Number of interactions to plot}

\item{res_dir}{Directory to save the plots}

\item{plot.velo}{Whether to plot the velocity}
}
\value{
A plot if return.plot is TRUE. Otherwise, plots are saved to the specified directory.
}
\description{
Plot 3D LR-velo ranked by the user-specified metric
}
