% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intrDEG.R
\name{print.csIntrDEG}
\alias{print.csIntrDEG}
\title{Print method for csIntrDEG objects}
\usage{
\method{print}{csIntrDEG}(x, ...)
}
\arguments{
\item{x}{A \code{csIntrDEG} object.}

\item{...}{Additional arguments (not used).}
}
\value{
Invisible NULL value.
}
\description{
Print method for csIntrDEG objects
}
