% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_gene.R
\name{print.CytosignalIntrDEG}
\alias{print.CytosignalIntrDEG}
\title{Print the CytosignalIntrDEG object representation to screen}
\usage{
\method{print}{CytosignalIntrDEG}(x, ...)
}
\arguments{
\item{x}{A \code{CytosignalIntrDEG} object.}

\item{...}{Passed to other methods}
}
\description{
Print the CytosignalIntrDEG object representation to screen
}
