% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.r
\name{purgeBeforeSave}
\alias{purgeBeforeSave}
\title{Remove imputed data and normalized imputed data from CytoSignal object to save disk space}
\usage{
purgeBeforeSave(object, purge.raw = T, purge.null = F)
}
\arguments{
\item{object}{CytoSignal object}
}
\value{
a CytoSignal object
}
\description{
Remove imputed data and normalized imputed data from CytoSignal object to save disk space
}
