% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intrDEG.R
\name{refineIntrDEG}
\alias{refineIntrDEG}
\title{Refine interaction-associated DEG with elastic net regression}
\usage{
refineIntrDEG(
  object,
  intrDEG,
  minSpot = 50,
  effectSizeThresh = 1,
  alphaTest = seq(0.5, 1, 0.1),
  nFold = 5,
  seed = NULL
)
}
\arguments{
\item{object}{A \code{\linkS4class{cytosignal2}} object.}

\item{intrDEG}{A \code{csIntrDEG} object returned by
\code{\link{findIntrDEG}}.}

\item{minSpot}{Minimum number of spots a gene must be expressed in to be
considered for modeling. Default \code{50}.}

\item{effectSizeThresh}{Minimum effect size a DEG must have to be considered
for modeling. Default \code{1}.}

\item{alphaTest}{A numeric vector of alpha values to test in the elastic
net regression, the model with the lowest loss will be considered. Default
\code{seq(0.5, 1, 0.1)} (i.e. 0.5, 0.6, 0.7, ..., 1).}

\item{nFold}{Number of folds to use in cross-validation. Default \code{5}.}

\item{seed}{Random seed for reproducibility. Default \code{NULL} runs under
user global random states.}
}
\value{
A \code{csIntrDEG} object with refined DEG lists.
}
\description{
After detecting DEGs associated with interaction significance using
\code{\link{findIntrDEG}}, this function further refines the DEG list
by fitting an elastic net regression model per interaction. The further
model refines the LRScore by predicting with the model using the expression
of the detected DEGs and the cluster labels. The genes that are selected by
the model are kept as the final DEG list.
}
\seealso{
\code{\link{findIntrDEG}}, \code{\link{plotIntrDEGHeatmap}}
}
