% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.r
\name{removeLowQuality}
\alias{removeLowQuality}
\title{Remove low quality cells and genes from raw counts}
\usage{
removeLowQuality(object, counts.thresh = 300, gene.thresh = 50)
}
\arguments{
\item{object}{CytoSignal object}

\item{counts.thresh}{threshold for cell counts}

\item{gene.thresh}{threshold for gene counts}
}
\value{
a CytoSignal object
}
\description{
Remove low quality cells and genes from raw counts
}
