% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REVIGO.R
\name{revigo}
\alias{revigo}
\title{Submit query to REVIGO and get result to R}
\usage{
revigo(
  GOterms,
  cutoff = 0.7,
  value = NULL,
  valueType = c("PValue", "Higher", "Lower", "HigherAbsolute", "HigherAbsLog2"),
  speciesTaxon = "0",
  measure = c("SIMREL", "LIN", "RESNIK", "JIANG"),
  removeObsolete = TRUE
)
}
\arguments{
\item{GOterms}{Character vector of GO term IDs.}

\item{cutoff}{Size of the result, between 0.4 to 0.9. Larger value returns
more result. Default \code{0.7}.}

\item{value}{A vector of additional information for each GO term. Default
\code{NULL} does not include additional information. Type of the value has
to be specified with \code{valueType}.}

\item{valueType}{Type of \code{value}. Default \code{"PValue"}. Optionally,
\code{"Higher"}, \code{"Lower"}, \code{"HigherAbsolute"},
\code{"HigherAbsLog2"}.}

\item{speciesTaxon}{Taxon ID of the species. Default \code{"0"} queries to
all available species. \code{"10090"} is for mouse and \code{"9606"} is for
human.}

\item{measure}{The semantic similarity measure to use. Default
\code{"SIMREL"}. Optionally, \code{"LIN"}, \code{"RESNIK"}, \code{"JIANG"}.}

\item{removeObsolete}{Logical. Whether to remove obsolete GO terms. Default
\code{TRUE}.}
}
\value{
A data.frame with many information
}
\description{
This function gets an input of vector of GO terms and submit query to REVIGO
via HTTP, and retrieves the result from the job.
The code is based on example provided by REVIGO, originally located at
http://revigo.irb.hr/FAQ

Please visit http://revigo.irb.hr/ for the web-based interactive tool, and
please see https://doi.org/10.1371/journal.pone.0021800 for detailed
introduction to the functionality.
}
\examples{
rvg <- revigo(c("GO:0003002", "GO:0007389", "GO:0007423", "GO:0009653"),
              speciesTaxon = 10090)
}
