% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nebula.R
\name{runNEBULA}
\alias{runNEBULA}
\title{Run NEBULA's association analysis to compare datasets within mergedCytoSignal object}
\usage{
runNEBULA(object, covariates, ncore = 1L, verbose = TRUE)
}
\arguments{
\item{object}{A \linkS4class{mergedCytoSignal} object}

\item{covariates}{A character vector of metadata covariates to include in the
model.}

\item{ncore}{Integer, number of cores to use for parallel computation.
Default \code{1L}.}

\item{verbose}{Logical, print progress messages. Default \code{TRUE}.}
}
\value{
Input \code{object} with slot \code{diff.results} and
\code{cont.results} updated with the results from NEBULA analysis.
}
\description{
After merging multiple CytoSignal objects and adding metadata covariates as
needed for modeling the comparison, this function runs NEBULA's association
analysis to identify significant interactions between cells from different
datasets and of different covariate levels.
}
\seealso{
\code{\link[=mergeCytoSignal]{mergeCytoSignal()}}
}
