% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{runPears.std}
\alias{runPears.std}
\title{Identify spatially significant interactions using std-corrected pearson correlation
Normal Moran's I test is not applicable here since the total number of the cell is too large, causing
unnacceptable computation cost. Here we use a modified version of Moran's I test, which is to take only
the top KNNs to compute the Moran's I test.}
\usage{
runPears.std(object, k = 10, weight = 2, score.slot = NULL)
}
\arguments{
\item{object}{A Cytosignal object}

\item{k}{The number of nearest neighbors to use}

\item{weight}{The weight of the nearest neighbors}

\item{score.slot}{Which LR score to use. Use the name specified with \code{tag}
when running \code{\link{inferLRScore}}.}
}
\value{
A Cytosignal object
}
\description{
Identify spatially significant interactions using std-corrected pearson correlation
Normal Moran's I test is not applicable here since the total number of the cell is too large, causing
unnacceptable computation cost. Here we use a modified version of Moran's I test, which is to take only
the top KNNs to compute the Moran's I test.
}
