% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object2.R
\name{setCluster}
\alias{setCluster}
\title{Set default cluster variable}
\usage{
setCluster(object, cluster)
}
\arguments{
\item{object}{A \linkS4class{cytosignal2} object.}

\item{cluster}{A single string for selecting a metadata column, or a vector
or factor for cluster assignment.}
}
\value{
The input \linkS4class{cytosignal2} object with updated parameters
at \code{object@parameters$cluster}.
}
\description{
Set an existing metadata variable as the default cluster annotation, or add
a new vector/factor as cluster annotation. This variable will be used by
default in cluster-wise inference and visualization functions.
}
