% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object2.R
\name{setDefaultColorBy}
\alias{setDefaultColorBy}
\title{Set default colorBy parameter}
\usage{
setDefaultColorBy(object, colorBy)
}
\arguments{
\item{object}{A \linkS4class{cytosignal2} object.}

\item{colorBy}{A character string specifying the metadata variable to set as
the default coloring variable.}
}
\value{
The input \linkS4class{cytosignal2} object with updated parameters
in \code{object@parameters}.
\itemize{
\item{\code{defaultColorBy}: The metadata variable set as the default
coloring variable.}
}
}
\description{
Set an existing metadata variable as the default coloring variable. This
variable will be used by default in visualization functions.
}
