% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor.R
\name{setParams}
\alias{setParams}
\title{Set parameters for correctly scaling the spatial coordinates}
\usage{
setParams(
  object,
  micronPerUnit,
  ballRadiusMicron = 200,
  thresh = 0.001,
  plot = TRUE
)
}
\arguments{
\item{object}{A \linkS4class{cytosignal2} object.}

\item{micronPerUnit}{Numeric, 1 spatial coordinate unit equals to how many
microns.}

\item{ballRadiusMicron}{Numeric, radius of the Epsilon ball in microns.
Default \code{200}.}

\item{thresh}{Numeric, threshold for the Gaussian kernel. Default
\code{0.001}.}

\item{plot}{Logical, whether to plot the scale bar after setting the
parameters. Default \code{TRUE}.}
}
\value{
The input \linkS4class{cytosignal2} object with updated parameters
in \code{object@parameters}.
\itemize{
\item{\code{micronPerUnit}: 1 spatial coordinate unit equals to how many
microns.}
\item{\code{ballRadiusMicron}: radius of the Epsilon ball in microns.}
\item{\code{ballRadiusCoord}: radius of the Epsilon ball in spatial
coordinate units.}
\item{\code{sigma}: sigma used for Gaussian kernel in modeling the
diffusion-dependent neighborhood.}
}
}
\description{
Set parameters required for modeling the neighborhood of LR interactions. A
plot is by default shown to help users to visually check if the parameters
look appropriate.
}
