% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.r
\name{show,CytoSignal-method}
\alias{show,CytoSignal-method}
\title{show method for CytoSignal}
\usage{
\S4method{show}{CytoSignal}(object)
}
\arguments{
\item{object}{CytoSignal object}
}
\description{
show method for CytoSignal
}
