% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.r
\name{show,lrScores-method}
\alias{show,lrScores-method}
\title{show method for lrScores object}
\usage{
\S4method{show}{lrScores}(object)
}
\arguments{
\item{object}{lrScores object}
}
\description{
show method for lrScores object
}
