% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nebula.R
\name{show,mergedCytoSignal-method}
\alias{show,mergedCytoSignal-method}
\title{Show method for mergedCytoSignal object}
\usage{
\S4method{show}{mergedCytoSignal}(object)
}
\arguments{
\item{object}{A mergedCytoSignal object}
}
\value{
Show a message with object summary, no return value
}
\description{
Show method for mergedCytoSignal object
}
