% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nebula.R
\name{showCov}
\alias{showCov}
\title{Show available covariate levels used in NEBULA analysis}
\usage{
showCov(object, intr.type = NULL)
}
\arguments{
\item{object}{A \linkS4class{mergedCytoSignal} object}

\item{intr.type}{A character string specifying the type of interaction to
show covariate levels for. Choose from \code{"diff"} for diffusion-dependent
or \code{"cont"} for contact-dependent interactions. Default \code{NULL}
returns the covariate levels used for both if they are the same, which is
true most of the time.}
}
\value{
A character vector
}
\description{
Show available covariate levels used in NEBULA analysis
}
