% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.r
\name{showImp}
\alias{showImp}
\title{show method for ImpData}
\usage{
showImp(object, slot.use = NULL)
}
\arguments{
\item{object}{CytoSignal object}

\item{slot.use}{slot to use}
}
\description{
show method for ImpData
}
