% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.r
\name{showIntr}
\alias{showIntr}
\title{show method for CytoSignal}
\usage{
showIntr(object, slot.use = NULL, signif.use = NULL)
}
\arguments{
\item{object}{CytoSignal object}

\item{slot.use}{slot to use}

\item{signif.use}{Significance level to use for plotting}
}
\description{
show method for CytoSignal
}
