% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.r
\name{showScore}
\alias{showScore}
\title{show method for lrScores}
\usage{
showScore(object, slot.use = NULL)
}
\arguments{
\item{object}{CytoSignal object}

\item{slot.use}{slot to use}
}
\description{
show method for lrScores
}
