% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.r
\name{showUnpt}
\alias{showUnpt}
\title{show intr.data in ImpData}
\usage{
showUnpt(object, slot.use = NULL)
}
\arguments{
\item{object}{CytoSignal object}

\item{slot.use}{slot to use}
}
\description{
show intr.data in ImpData
}
