% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.r
\name{showVelo}
\alias{showVelo}
\title{show method for lrVelo}
\usage{
showVelo(object, slot.use = NULL)
}
\arguments{
\item{object}{CytoSignal object}

\item{slot.use}{slot to use}
}
\description{
show method for lrVelo
}
