% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object2.R
\name{subset.cytosignal2}
\alias{subset.cytosignal2}
\title{Subsetting cytosignal2 object by spots}
\usage{
\method{subset}{cytosignal2}(x, subset, drop = FALSE, keepNeighbor = FALSE, ...)
}
\arguments{
\item{x}{A \linkS4class{cytosignal2} object.}

\item{subset}{Logical, numeric, or character vector specifying the spots to
keep. Usually passed with expression using metadata variables.}

\item{drop}{Logical, whether to drop unused factor levels in metadata after
subsetting. Default \code{FALSE}.}

\item{keepNeighbor}{Logical, whether to keep the subset neighbor graphs
regardless of whether the subset spots still have the same neighbor structure.
Default \code{FALSE} remove graphs from the subset object.}

\item{...}{Additional arguments (not used).}
}
\value{
A \linkS4class{cytosignal2} object subsetted by spots.
}
\description{
Subset a \linkS4class{cytosignal2} object by spots.
}
