% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nebula.R
\name{subset.mergedCytoSignal}
\alias{subset.mergedCytoSignal}
\title{Subset a mergedCytoSignal obejct}
\usage{
\method{subset}{mergedCytoSignal}(x, subset, drop = FALSE, ...)
}
\arguments{
\item{x}{A \linkS4class{mergedCytoSignal} object}

\item{subset}{logical expression indicating the cell/spots to keep.}

\item{drop}{passed on to \code{[} indexing operator.}

\item{...}{Not used}
}
\value{
A \linkS4class{mergedCytoSignal} object with metadata and LRscores
for only selected cells/spots. All test results and modeling setup are
cleaned.
}
\description{
Subset a \linkS4class{mergedCytoSignal} object by cells/spots. Mainly for
conducting analyses over a subset of datasets within the current object.
}
\examples{
\dontrun{
submultics <- subset(
    x = multics,
    subset = multics$dataset \%in\% c("disease1", "disease2"),
    drop = TRUE
)
}
}
