% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{subsetCSDB}
\alias{subsetCSDB}
\title{Subset cytosignal database object by genes}
\usage{
subsetCSDB(
  csdb,
  ligandGenes = NULL,
  receptorGenes = NULL,
  type = NULL,
  LorR = FALSE
)
}
\arguments{
\item{csdb}{A \code{\link{csdb-class}} object.}

\item{ligandGenes}{A character vector of regular expression to match ligand
genes to keep. Complexes are kept if any of their member genes match. Default
\code{NULL} skip the filtering by ligand genes.}

\item{receptorGenes}{A character vector of regular expression to match
receptor genes to keep. Complexes are kept if any of their member genes
match. Default \code{NULL} skip the filtering by receptor genes.}

\item{type}{Type of interaction to keep. Either \code{'diffusion'} or
\code{'contact'}. Default \code{NULL} skip the filtering by type.}

\item{LorR}{Logical, whether to keep interactions that match the ligand gene
pattern or the receptor gene pattern. Default \code{FALSE} keeps interactions
that match both patterns at the same time.}
}
\value{
The subset \code{csdb}.
}
\description{
This function allows user to limit the interaction database to only those
involving specified ligand and/or receptor genes, or type
(diffusion-/contact-dependent).
}
