% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.r
\name{suggestInterval}
\alias{suggestInterval}
\title{Suggest cell intervels for scaling}
\usage{
suggestInterval(object)
}
\arguments{
\item{object}{CytoSignal object}
}
\value{
a vector of intervels
}
\description{
This function is used to estimate the actual cell intervel in cells.loc after scaling
returns the top 5 least intervels, users can choose to which to use.
}
