% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.r
\name{suggestScaleFactor}
\alias{suggestScaleFactor}
\title{Suggest scaling factor of real units to spatial units}
\usage{
suggestScaleFactor(object, cell.intervel = NULL)
}
\arguments{
\item{object}{CytoSignal object}

\item{cell.intervel}{Distance between two cell in physical units}
}
\value{
a vector of possible scaling factors
}
\description{
This function is used to estimate the scaling facor of real units to spatial units.
Returns the top 5 possible scaling factors, users can choose to which to use.
}
