% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utility.r
\name{uniqueColors}
\alias{uniqueColors}
\title{Generate character vector of unique color hex codes
Base color palette adopts \code{ggsci:::ggsci_db$igv$default}}
\usage{
uniqueColors(n)
}
\arguments{
\item{n}{Number of unique colors to generate. When not exceeding existing
number of colors, just return the first \code{n} colors, otherwise generate
interpolated colors.}
}
\description{
Generate character vector of unique color hex codes
Base color palette adopts \code{ggsci:::ggsci_db$igv$default}
}
