% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRvelo.r
\name{veloNicheLR.CytoSignal}
\alias{veloNicheLR.CytoSignal}
\title{Sub function for veloNicheLR, input a CytoSignal object}
\usage{
\method{veloNicheLR}{CytoSignal}(object, lig.slot, recep.slot, intr.db.name, tag = NULL)
}
\arguments{
\item{object}{A Cytosignal object}

\item{lig.slot}{The ligand slot to use}

\item{recep.slot}{The receptor slot to use}

\item{intr.db.name}{The intr database name to use}

\item{nn.use}{slot that the neighbor index should be taken from, by default is the same as
the recep.slot. For example, if velo.obj = GauEps-DT, then nn.use = "DT".
nn.use could also be a user-defind factor.}
}
\value{
A Cytosignal object
}
\description{
Sub function for veloNicheLR, input a CytoSignal object
}
