% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{veloNicheLR}
\alias{veloNicheLR}
\title{Compute the LR velo for specific ligand-receptor imputation obj pairs}
\usage{
veloNicheLR(object, ...)
}
\arguments{
\item{object}{A Cytosignal object}

\item{lig.slot}{The ligand slot to use}

\item{recep.slot}{The receptor slot to use}

\item{intr.db.name}{The intr database name to use}

\item{nn.use}{The neighbor index as niche}
}
\value{
A Cytosignal object
}
\description{
Compute the LR velo for specific ligand-receptor imputation obj pairs
}
