% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.r
\name{changeUniprot.CytoSignal}
\alias{changeUniprot.CytoSignal}
\title{Sub function for changeUniprot, input a cytosignal object}
\usage{
\method{changeUniprot}{CytoSignal}(object, verbose = T)
}
\arguments{
\item{object}{A Cytosignal object}

\item{verbose}{Whether to print out the progress}
}
\value{
A Cytosignal object
}
\description{
Sub function for changeUniprot, input a cytosignal object
}
