% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.r
\name{changeUniprot}
\alias{changeUniprot}
\title{Subset the imputed data (intr.data) by the specified intr index}
\usage{
changeUniprot(object, ...)
}
\arguments{
\item{object}{A Cytosignal object}

\item{intr.index}{The intr index to use for subsetting}

\item{slot.use}{The slot to use for subsetting}
}
\value{
A Cytosignal object
}
\description{
Subset the imputed data (intr.data) by the specified intr index
}
