% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{csToH5AD}
\alias{csToH5AD}
\alias{H5ADToCS}
\title{Conversion between CytoSignal object and AnnData object in H5AD file}
\usage{
csToH5AD(object, filename, overwrite = FALSE)

H5ADToCS(filename, layer = "X")
}
\arguments{
\item{object}{A CytoSignal object.}

\item{filename}{Name of the H5AD file.}

\item{overwrite}{Whether to overwrite the existing H5AD file. Default
\code{FALSE}.}

\item{layer}{Data from the H5AD file to extract. Default \code{"X"} extracts
\code{adata.X}. Use \code{"raw/X"} to extract \code{adata.raw.X}. Use
\code{"layers/key"} to extract layer data with key \code{"key"}.}
}
\value{
\code{csToH5AD} writes an H5AD file to disk and returns \code{NULL}.
\code{H5ADToCS} returns a CytoSignal object.
}
\description{
Extract data from H5AD file to a CytoSignal object or write a CytoSignal
object to a H5AD file.
}
