% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.r
\docType{class}
\name{CytoSignal-class}
\alias{CytoSignal-class}
\alias{CytoSignal}
\title{The CytoSignal Class}
\description{
The CytoSignal object is created from one spatial transcriptomic dataset. A
CytoSignal object can be created with \code{\link{createCytoSignal}}.
}
\section{Slots}{

\describe{
\item{\code{counts}}{Raw gene expression count matrix, sparse dgCMatrix, gene x cell.}

\item{\code{cell.loc}}{Matrix of cell locations, cell by axis.}

\item{\code{clusters}}{cluster assignments, named factor.}

\item{\code{imputation}}{List of imputation objects.}

\item{\code{lrscore}}{List of lrScore objects.}

\item{\code{velo}}{List of velo objects.}

\item{\code{intr.valid}}{List of interaction database. Added with
\code{\link{addIntrDB}}.}

\item{\code{cell.data}}{data.frame of cell data.}

\item{\code{parameters}}{List of parameters for gaussian imputation}

\item{\code{log}}{Log of each main steps}

\item{\code{version}}{Version of package used to create object. Should not be
modified by users.}
}}

