% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{findNNDT}
\alias{findNNDT}
\title{Find the direct connected neighbor of each cell, using Delaunay triangulation}
\usage{
findNNDT(object, max.r = NULL, weight.sum = 2, imp.name = "DT")
}
\arguments{
\item{object}{A CytoSignal object or a matrix of cell location.}

\item{max.r}{The maximum radius of the edges, by default is the r.diffuse.scale,
or the value inferred by \code{\link{inferEpsParams}}.}

\item{weight.sum}{The sum of the weights}

\item{imp.name}{The name of the analysis.}
}
\value{
Returns the original object with result updated in "imputation" slot.
An "ImpData" object is created with its "nn.graph" slot filled a NxN sparse
matrix indicating the NN relationship and weights. A column is considered as
a center of the epsilon ball and the rows containing non-zero values are
its neighbors.
}
\description{
Find the direct connected neighbors of each cell, and assigning an averaged
weight to each neighbor.
}
