% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{findNNGauEB.CytoSignal}
\alias{findNNGauEB.CytoSignal}
\title{Sub function for findNNGauEB, input a Cytosignal object}
\usage{
\method{findNNGauEB}{CytoSignal}(object, eps = NULL, sigma = NULL, weight.sum = 2, tag = NULL)
}
\arguments{
\item{object}{A Cytosignal object}

\item{eps}{The radius of the epsilon ball}

\item{sigma}{The sigma of the Gaussian kernel}

\item{weight.sum}{The sum of the weights}
}
\value{
A Cytosignal object
}
\description{
Sub function for findNNGauEB, input a Cytosignal object
}
