% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{findNNGauEB}
\alias{findNNGauEB}
\title{Find the neighbors of each cell in the epsilon ball}
\usage{
findNNGauEB(object, ...)
}
\arguments{
\item{object}{A Cytosignal object}

\item{eps}{The radius of the epsilon ball}

\item{sigma}{The sigma of the Gaussian kernel}

\item{self.weight}{weight of the index cell}
}
\value{
A list of neighbors and their distances
}
\description{
Find the neighbors of each cell in the epsilon ball
}
