% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{findNNRaw}
\alias{findNNRaw}
\title{Create a ImpData object using raw data without imputation}
\usage{
findNNRaw(object)
}
\arguments{
\item{object}{A Cytosignal object}
}
\value{
A Cytosignal object
}
\description{
Create a ImpData object using raw data without imputation
}
