% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrscore2.R
\name{getTopIntr}
\alias{getTopIntr}
\title{Get interaction table ranked by significance}
\usage{
getTopIntr(
  object,
  alpha = 0.05,
  minExp = 100,
  minSpot = 100,
  type = NULL,
  topN = NULL,
  balanceType = TRUE
)
}
\arguments{
\item{object}{A \code{\linkS4class{cytosignal2}} object with LRscore
inference done with \code{\link{inferLRScore}}.}

\item{alpha}{Numeric, significance level to call an LRscore significant in
a spot. Default is \code{0.05}.}

\item{minExp}{Integer, minimum number of spots with non-zero expression for
each member gene of an interaction. Interactions with any member gene having
less than \code{minExp} spots with non-zero expression will be filtered out.
Default is \code{100}.}

\item{minSpot}{Integer, minimum number of spots with significant LRscore
for an interaction to be kept. Default is \code{100}.}

\item{type}{Character vector, interaction types to keep. Options are
\code{"diffusion"} and \code{"contact"}. Default \code{NULL} ranks both types
together.}

\item{topN}{Integer, number of top interactions to return. Default
\code{NULL} keeps all high-quality interactions.}

\item{balanceType}{Logical, effective when \code{type = NULL}. Whether to
take \code{topN} interactions from each type. Default is \code{TRUE}.}
}
\value{
A data frame of class \code{csdb}, the filtered and ranked
interaction table with an additional column \code{n_signif_spots} indicating
the number of spots with significant LRscore for each interaction.
}
\description{
This function counts the number of spots with significant LRscores for each
interaction and rank the interactions by the counts. It also filters out
low quality interactions based on member gene expression.
}
