% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{imputeNiche}
\alias{imputeNiche}
\title{Impute the data using the specified method}
\usage{
imputeNiche(object, ...)
}
\arguments{
\item{object}{A Cytosignal object}

\item{...}{Other parameters}

\item{method}{The method to use for imputation}
}
\value{
A Cytosignal object
}
\description{
Impute the data using the specified method
}
