% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{imputeNiche}
\alias{imputeNiche}
\title{Impute the data}
\usage{
imputeNiche(
  object,
  imp.use = NULL,
  weights = c("none", "mean", "counts", "dist")
)
}
\arguments{
\item{object}{A Cytosignal object with \code{findNN*} performed.}

\item{imp.use}{The type of neighbors. The \code{imp.name} used when finding
neighbors. Default \code{NULL} uses the last NN type computed.}

\item{weights}{Weighting strategy. Default \code{"none"} use weights computed
in \code{findNN*}. Alternatively, choose from \code{"mean"}, \code{"counts"},
\code{"dist"}.}
}
\value{
Returns the original object with imputation result updated. For the
selected "ImpData" specified by \code{imp.use}, the following slots are
updated:
\itemize{
\item{"imp.data"}{The imputed data.}
\item{"nn.graph"}{The neighbor graph, updated if \code{"weights"} is not
"none".}
\item{"scale.fac"}{The scale factor of the imputed data.}
}
}
\description{
Impute the quantitative value of ligand received or receptor expressed at
each bead basing on the NN weighting previously determined with
\code{findNN*} functions.
}
