% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{inferEpsParams}
\alias{inferEpsParams}
\title{Infer the parameters of the Gaussian kernel}
\usage{
inferEpsParams(
  object,
  tech.res,
  bin_size,
  loc.d,
  r.eps.real = 200,
  thresh = 0.001
)
}
\arguments{
\item{object}{A Cytosignal object}

\item{tech.res}{The resolution of the technology}

\item{bin_size}{The size of the bin}

\item{loc.d}{The scaled distance between two cells}

\item{r.eps.real}{The radius of the epsilon ball in tech resolution in um, default 200 um}

\item{thresh}{The total signal out of the epsilon ball}
}
\value{
A Cytosignal object
}
\description{
Infer the parameters of the Gaussian kernel
}
