% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrscore2.R
\name{inferLRScore}
\alias{inferLRScore}
\title{Infer the LR score with significance tested}
\usage{
inferLRScore(object, smoothR = FALSE, permSize = 1e+05)
}
\arguments{
\item{object}{A \code{\linkS4class{cytosignal2}} object, with neighbor graphs
constructed.}

\item{smoothR}{Logical, whether to smooth the receptor expression using the
contact-dependent neighbor graph. Default is \code{FALSE}.}

\item{permSize}{Integer, sample size to reach in the permutation test.
Default \code{1e5}.}
}
\value{
A \code{\linkS4class{cytosignal2}} object with LR scores inferred and
stored in the \code{lrScore} slot.
}
\description{
This function imputes the niche expression of potential ligand and receptor
genes using the neighbor graphs, and computes the final LR score. Optionally,
with \code{smoothR = TRUE}, this function allows smoothing the receptor
expression of a spot using the neighbor graph of contact-dependent
neighborhood. This is useful when the data is of high spatial resolution
and/or high sparsity. This function further performs a permutation test to
infer the significance of each LR interaction in each spot.
}
