% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{inferSignif}
\alias{inferSignif}
\title{Infer significance of LR scores}
\usage{
inferSignif(
  object,
  fdr.method = c("spatialFDR", "fdr"),
  p.value = 0.05,
  reads.thresh = 100,
  sig.thresh = 100,
  slot.use = NULL,
  nn.use = NULL
)
}
\arguments{
\item{object}{A Cytosignal object}

\item{fdr.method}{The false discovery rate method to use. Choose from
\code{"spatialFDR"} and \code{"fdr"}. Default \code{"spatialFDR"}.}

\item{p.value}{A numeric scalar. The p-value threshold to use for filtering
significant interactions. Default \code{0.05}.}

\item{reads.thresh}{A numeric scalar. The minimum number of reads for a
ligand-receptor interaction to be considered. Default \code{100}.}

\item{sig.thresh}{A numeric scalar. The minimum number of of beads for a
ligand-receptor interaction to be considered. Default \code{100}.}

\item{slot.use}{Which LR score to use. Use the name specified with \code{tag}
when running \code{\link{inferIntrScore}}. Default \code{NULL} apply
specified significance and filtering criteria to all available LRscore slots.}

\item{nn.use}{Which imputation to use. Default the imputation used for
deriving the LRScore specified with \code{slot.use}. Use the name specified
with \code{tag} when running \code{\link{findNNGauEB}}; use \code{"DT"} for
imputation produced with \code{\link{findNNDT}}; or use \code{"Raw"} for
imputation produced with \code{\link{findNNRaw}}.}
}
\value{
A Cytosignal object
}
\description{
Infer significance of LR scores
}
