% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{inferSignif}
\alias{inferSignif}
\title{Infer significance of LR scores}
\usage{
inferSignif(
  object,
  correctBy = c("cell", "intr"),
  p.value = 0.05,
  reads.thresh = 100,
  sig.thresh = 100,
  slot.use = NULL,
  nn.use = NULL
)
}
\arguments{
\item{object}{A Cytosignal object}

\item{correctBy}{Spatial FDR correction method. Choose from \code{"cell"} or
\code{"intr"}.}

\item{p.value}{A numeric value of p-value threshold}

\item{reads.thresh}{The minimum number of reads for a ligand-receptor
interaction to be considered.}

\item{sig.thresh}{The minimum number of beads for a ligand-receptor
interaction to be considered.}

\item{slot.use}{Which LR score to use. Use the name specified with \code{tag}
when running \code{\link{inferLRScore}}.}

\item{nn.use}{Which imputation to use. Default the imputation used for
deriving the LRScore specified with \code{slot.use}. Use the name specified
with \code{tag} when running \code{\link{findNNGauEB}}; use \code{"DT"} for
imputation produced with \code{\link{findNNDT}}; or use \code{"Raw"} for
imputation produced with \code{\link{findNNRaw}}.}
}
\value{
A Cytosignal object
}
\description{
Infer significance of LR scores
}
