% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{inferSignif}
\alias{inferSignif}
\title{Infer significance of LR scores}
\usage{
inferSignif(object, ...)
}
\arguments{
\item{object}{A Cytosignal object}

\item{...}{Other arguments}

\item{lrscore.mtx}{A matrix of LR scores}

\item{null.lrscore.mtx}{A matrix of NULL LR scores}

\item{nb.fac}{A factor of nearest neighbors index}

\item{intr.db}{A interaction database}

\item{gene_to_uniprot}{A dataframe of gene to uniprot mapping}

\item{p.thresh}{A numeric value of p-value threshold}

\item{reads.thresh}{A numeric value of reads threshold}

\item{sig.thresh}{A numeric value of significance threshold}
}
\value{
A list of indexes of significant cells
}
\description{
Infer significance of LR scores
}
