% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nebula.R
\name{mergeCytoSignal}
\alias{mergeCytoSignal}
\title{Create mergedCytoSignal object}
\usage{
mergeCytoSignal(objList, metadata = NULL, name.by = NULL)
}
\arguments{
\item{objList}{A list of CytoSignal object.}

\item{metadata}{A data.frame object. Default \code{NULL}. If provided, must
have as many rows as the length of \code{objList} with rows in matching
order to \code{objList}.}

\item{name.by}{A column name in \code{metadata} indicating where to find the
dataset names. Default \code{NULL} names datasets by "dataset1", "dataset2"
and etc, and will create column "dataset" in the metadata.}
}
\value{
A \linkS4class{mergedCytoSignal} object preprocessed.
}
\description{
This function accepts a list of CytoSignal objects together with a metadata
\code{data.frame} that provides dataset level metadata covariates, which will
be required for the cross-dataset analysis.

The function will preprocess each CytoSignal object with cross-dataset
analysis specific steps, including imputing diffusible LR values using
Gaussion Epsilon model with weight 1.
}
