% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nebula.R
\docType{class}
\name{mergedCytoSignal-class}
\alias{mergedCytoSignal-class}
\title{mergedCytoSignal Class for cross-dataset analysis of spatial cell-cell
communication}
\description{
mergedCytoSignal Class for cross-dataset analysis of spatial cell-cell
communication
}
\section{Slots}{

\describe{
\item{\code{lrscore}}{A sparse matrix (dgCMatrix class) with LR scores from all
dataset merged into one.}

\item{\code{metadata}}{A data.frame with location level metadata covariates.}
}}

