% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{normCounts}
\alias{normCounts}
\title{Normalize the data using the specified method}
\usage{
normCounts(object, ...)
}
\arguments{
\item{object}{A Cytosignal object}

\item{...}{Other parameters}

\item{method}{The method to use for normalization}
}
\value{
A Cytosignal object
}
\description{
Normalize the data using the specified method
}
