% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRscores.r
\name{permuteLR}
\alias{permuteLR}
\title{Permute Imputation Results of specific imputation method}
\usage{
permuteLR(object, slot.use = NULL, norm.method = "default", perm.size = 1e+05)
}
\arguments{
\item{object}{A CytoSignal object with \code{\link{inferLRScore}} run
beforehand.}

\item{slot.use}{Which LR score to use. Use the name specified with \code{tag}
when running \code{\link{inferLRScore}}.}

\item{perm.size}{Size of the permutation test. Defult \code{100000}.}

\item{norm.me}{Normalization method. See \code{\link{normCounts}} for detail.}
}
\value{
A CytoSignal object
}
\description{
This function permutes the imputation methods from which a given LR score is
calculated. Note that all rounds of permutation will use the same shuffle and
sample index.
}
