% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor.R
\name{plotNeighbor}
\alias{plotNeighbor}
\alias{plotNeighborDiff}
\alias{plotNeighborCont}
\title{Visualize neighbor weights of a spot}
\usage{
plotNeighbor(object, type = c("diffusion", "contact"), index = NULL)

plotNeighborDiff(object, index = NULL)

plotNeighborCont(object, index = NULL)
}
\arguments{
\item{object}{A \code{cytosignal2} object. With \code{\link{findNeighbor}} called.}

\item{type}{Type of neighborhood graph to visualize. Either
\code{'diffusion'} or \code{'contact'}.}

\item{index}{Index of the spot to visualize. Default \code{NULL} selects
the spot with the most neighbors. Can be an integer index, character ID, or
logical selector vector.}
}
\value{
A ggplot2 object showing the spatial coordinates colored by the neighbor
weights of the selected spot.
}
\description{
This function makes a spatial coordinate plot colored by the weights of a
specified spot in the selected type of neighborhood graph. By default,
the spot with the most neighbors is selected.
}
\seealso{
\code{\link[=findNeighbor]{findNeighbor()}}
}
