% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{plotVelo}
\alias{plotVelo}
\title{Plot the lrvelo results}
\usage{
plotVelo(
  object,
  num.plot,
  plot_dir,
  plot.fmt,
  slot.use = NULL,
  use.rank = NULL,
  plot.clusters = TRUE,
  colors.list = NULL,
  z.scaler = 0.03,
  use.cex = 0.1,
  use.shape = 16,
  use_xbins = 15,
  use_ybins = 15,
  arrow.line.width = 0.6,
  arrow.width = 0.06,
  width = 3,
  height = 3,
  use.phi = 45,
  use.theta = -17
)
}
\description{
Plot the lrvelo results
}
